Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the substring function against a set of input data sizes:
input.sizes = 10^seq(1, 4, by = 0.5)
asymptoticTimings(substring(paste(rep("A", N), collapse = ""), 1:N, 1:N), input.sizes)

#>     Timings  Data sizes
#> 1     52400    10.00000
#> 2     13500    10.00000
#> 3     10100    10.00000
#> 4      9000    10.00000
#> 5      8800    10.00000
#> 6      9000    10.00000
#> 7      8900    10.00000
#> 8      8600    10.00000
#> 9      8400    10.00000
#> 10     8700    10.00000
#> 11     8500    10.00000
#> 12     8600    10.00000
#> 13     8700    10.00000
#> 14     8500    10.00000
#> 15     8600    10.00000
#> 16     8600    10.00000
#> 17     8600    10.00000
#> 18     8500    10.00000
#> 19     8600    10.00000
#> 20     9000    10.00000
#> 21     8700    10.00000
#> 22     8800    10.00000
#> 23     8600    10.00000
#> 24     8500    10.00000
#> 25     8600    10.00000
#> 26     8700    10.00000
#> 27     8800    10.00000
#> 28     8600    10.00000
#> 29    11700    10.00000
#> 30    12400    10.00000
#> 31     8500    10.00000
#> 32     8900    10.00000
#> 33     9100    10.00000
#> 34     8800    10.00000
#> 35     9100    10.00000
#> 36     8700    10.00000
#> 37     8700    10.00000
#> 38     8700    10.00000
#> 39     8900    10.00000
#> 40     8500    10.00000
#> 41     8900    10.00000
#> 42     8600    10.00000
#> 43     8600    10.00000
#> 44     8900    10.00000
#> 45     8700    10.00000
#> 46     8400    10.00000
#> 47     8600    10.00000
#> 48     8400    10.00000
#> 49     8700    10.00000
#> 50     9000    10.00000
#> 51     8400    10.00000
#> 52     8700    10.00000
#> 53     8800    10.00000
#> 54     8500    10.00000
#> 55     8800    10.00000
#> 56     8700    10.00000
#> 57     8500    10.00000
#> 58     8700    10.00000
#> 59     8800    10.00000
#> 60     8800    10.00000
#> 61     8600    10.00000
#> 62     8400    10.00000
#> 63     8600    10.00000
#> 64     8600    10.00000
#> 65     8600    10.00000
#> 66     8700    10.00000
#> 67     8600    10.00000
#> 68     8700    10.00000
#> 69     9000    10.00000
#> 70     8700    10.00000
#> 71     8600    10.00000
#> 72     8500    10.00000
#> 73     8700    10.00000
#> 74     8600    10.00000
#> 75     8600    10.00000
#> 76     8700    10.00000
#> 77     8700    10.00000
#> 78     8600    10.00000
#> 79     8600    10.00000
#> 80     8600    10.00000
#> 81     8500    10.00000
#> 82     8800    10.00000
#> 83     8600    10.00000
#> 84     9200    10.00000
#> 85     8800    10.00000
#> 86     8500    10.00000
#> 87     8700    10.00000
#> 88     8800    10.00000
#> 89     9000    10.00000
#> 90     8700    10.00000
#> 91     8700    10.00000
#> 92     8700    10.00000
#> 93     8900    10.00000
#> 94     8600    10.00000
#> 95     8600    10.00000
#> 96     8900    10.00000
#> 97     8800    10.00000
#> 98     8800    10.00000
#> 99     8700    10.00000
#> 100    8700    10.00000
#> 101   51300    31.62278
#> 102   17200    31.62278
#> 103   13700    31.62278
#> 104   12700    31.62278
#> 105   12400    31.62278
#> 106   13200    31.62278
#> 107   12300    31.62278
#> 108   12100    31.62278
#> 109   12400    31.62278
#> 110   12700    31.62278
#> 111   12400    31.62278
#> 112   12600    31.62278
#> 113   12200    31.62278
#> 114   12300    31.62278
#> 115   11800    31.62278
#> 116   12800    31.62278
#> 117   12400    31.62278
#> 118   12300    31.62278
#> 119   12100    31.62278
#> 120   12200    31.62278
#> 121   12100    31.62278
#> 122   12500    31.62278
#> 123   12500    31.62278
#> 124   12300    31.62278
#> 125   11900    31.62278
#> 126   12000    31.62278
#> 127   12400    31.62278
#> 128   12400    31.62278
#> 129   12300    31.62278
#> 130   12600    31.62278
#> 131   12200    31.62278
#> 132   12200    31.62278
#> 133   12300    31.62278
#> 134   12200    31.62278
#> 135   17700    31.62278
#> 136   13400    31.62278
#> 137   11800    31.62278
#> 138   11900    31.62278
#> 139   13200    31.62278
#> 140   12300    31.62278
#> 141   12200    31.62278
#> 142   11900    31.62278
#> 143   12700    31.62278
#> 144   12500    31.62278
#> 145   12400    31.62278
#> 146   12300    31.62278
#> 147   12200    31.62278
#> 148   12900    31.62278
#> 149   13200    31.62278
#> 150   12400    31.62278
#> 151   12400    31.62278
#> 152   12100    31.62278
#> 153   12500    31.62278
#> 154   12800    31.62278
#> 155   12300    31.62278
#> 156   11900    31.62278
#> 157   12000    31.62278
#> 158   13000    31.62278
#> 159   12400    31.62278
#> 160   12100    31.62278
#> 161   12200    31.62278
#> 162   12300    31.62278
#> 163   12000    31.62278
#> 164   12000    31.62278
#> 165   12100    31.62278
#> 166   11800    31.62278
#> 167   11900    31.62278
#> 168   12400    31.62278
#> 169   11800    31.62278
#> 170   12100    31.62278
#> 171   12000    31.62278
#> 172   12000    31.62278
#> 173   12400    31.62278
#> 174   12000    31.62278
#> 175   11900    31.62278
#> 176   14000    31.62278
#> 177   12900    31.62278
#> 178   14100    31.62278
#> 179   12400    31.62278
#> 180   12200    31.62278
#> 181   12100    31.62278
#> 182   12200    31.62278
#> 183   12100    31.62278
#> 184   13000    31.62278
#> 185   12000    31.62278
#> 186   11900    31.62278
#> 187   12000    31.62278
#> 188   12200    31.62278
#> 189   11900    31.62278
#> 190   13300    31.62278
#> 191   12100    31.62278
#> 192   12300    31.62278
#> 193   12500    31.62278
#> 194   12100    31.62278
#> 195   12600    31.62278
#> 196   12000    31.62278
#> 197   12000    31.62278
#> 198   12200    31.62278
#> 199   11800    31.62278
#> 200   12600    31.62278
#> 201   78200   100.00000
#> 202   29200   100.00000
#> 203   24500   100.00000
#> 204   23000   100.00000
#> 205   22600   100.00000
#> 206   23600   100.00000
#> 207   23600   100.00000
#> 208   23200   100.00000
#> 209   24700   100.00000
#> 210   23100   100.00000
#> 211   23000   100.00000
#> 212   22200   100.00000
#> 213   22100   100.00000
#> 214   22500   100.00000
#> 215   29400   100.00000
#> 216   67700   100.00000
#> 217   79200   100.00000
#> 218   65500   100.00000
#> 219   65200   100.00000
#> 220   64300   100.00000
#> 221   60900   100.00000
#> 222   63200   100.00000
#> 223   61000   100.00000
#> 224   67200   100.00000
#> 225   51500   100.00000
#> 226   65700   100.00000
#> 227   63600   100.00000
#> 228   61500   100.00000
#> 229   64600   100.00000
#> 230   65100   100.00000
#> 231   65200   100.00000
#> 232   62100   100.00000
#> 233   61800   100.00000
#> 234   64600   100.00000
#> 235   48000   100.00000
#> 236   49000   100.00000
#> 237   47800   100.00000
#> 238   42300   100.00000
#> 239   43400   100.00000
#> 240   47300   100.00000
#> 241   56100   100.00000
#> 242   45900   100.00000
#> 243   50300   100.00000
#> 244   60700   100.00000
#> 245   53300   100.00000
#> 246   57000   100.00000
#> 247   66300   100.00000
#> 248   64400   100.00000
#> 249   67300   100.00000
#> 250   66200   100.00000
#> 251   68700   100.00000
#> 252   69200   100.00000
#> 253   57100   100.00000
#> 254   66200   100.00000
#> 255   33800   100.00000
#> 256   47600   100.00000
#> 257   51900   100.00000
#> 258   62100   100.00000
#> 259   54800   100.00000
#> 260   47400   100.00000
#> 261   57900   100.00000
#> 262   53100   100.00000
#> 263   45600   100.00000
#> 264   47600   100.00000
#> 265   51100   100.00000
#> 266   64000   100.00000
#> 267   45400   100.00000
#> 268   60800   100.00000
#> 269   52600   100.00000
#> 270   50800   100.00000
#> 271   48800   100.00000
#> 272   45700   100.00000
#> 273   43500   100.00000
#> 274   42200   100.00000
#> 275   46100   100.00000
#> 276   43800   100.00000
#> 277   43800   100.00000
#> 278   67100   100.00000
#> 279   43400   100.00000
#> 280   27500   100.00000
#> 281   23200   100.00000
#> 282   22100   100.00000
#> 283   22200   100.00000
#> 284   22300   100.00000
#> 285   22200   100.00000
#> 286   22000   100.00000
#> 287   21700   100.00000
#> 288   22700   100.00000
#> 289   22300   100.00000
#> 290   22300   100.00000
#> 291   23400   100.00000
#> 292   22500   100.00000
#> 293   21500   100.00000
#> 294   22600   100.00000
#> 295   22300   100.00000
#> 296   22600   100.00000
#> 297   21900   100.00000
#> 298   22300   100.00000
#> 299   21800   100.00000
#> 300   21800   100.00000
#> 301  137300   316.22777
#> 302   68000   316.22777
#> 303   60300   316.22777
#> 304   59900   316.22777
#> 305   62800   316.22777
#> 306   66100   316.22777
#> 307   63900   316.22777
#> 308   62600   316.22777
#> 309   59500   316.22777
#> 310   66600   316.22777
#> 311   68300   316.22777
#> 312   76700   316.22777
#> 313   64800   316.22777
#> 314   65500   316.22777
#> 315   65000   316.22777
#> 316   65600   316.22777
#> 317   61600   316.22777
#> 318   74200   316.22777
#> 319  107600   316.22777
#> 320  105700   316.22777
#> 321   72100   316.22777
#> 322   68600   316.22777
#> 323   65400   316.22777
#> 324   64900   316.22777
#> 325   73400   316.22777
#> 326   99900   316.22777
#> 327   67500   316.22777
#> 328   64300   316.22777
#> 329   63100   316.22777
#> 330   62000   316.22777
#> 331   63500   316.22777
#> 332   93700   316.22777
#> 333   64100   316.22777
#> 334   63200   316.22777
#> 335   60300   316.22777
#> 336   65100   316.22777
#> 337   94700   316.22777
#> 338   71800   316.22777
#> 339   62000   316.22777
#> 340   72000   316.22777
#> 341   63300   316.22777
#> 342   64100   316.22777
#> 343   63600   316.22777
#> 344   63000   316.22777
#> 345   61500   316.22777
#> 346   61000   316.22777
#> 347   61100   316.22777
#> 348   62300   316.22777
#> 349   62500   316.22777
#> 350   58900   316.22777
#> 351   62400   316.22777
#> 352   61500   316.22777
#> 353   60100   316.22777
#> 354   58200   316.22777
#> 355   57500   316.22777
#> 356   94200   316.22777
#> 357   64400   316.22777
#> 358   57400   316.22777
#> 359   57100   316.22777
#> 360   57400   316.22777
#> 361   57700   316.22777
#> 362   59100   316.22777
#> 363   86400   316.22777
#> 364  126500   316.22777
#> 365  117700   316.22777
#> 366  103500   316.22777
#> 367  102000   316.22777
#> 368  108100   316.22777
#> 369  106200   316.22777
#> 370  101500   316.22777
#> 371  112700   316.22777
#> 372  122400   316.22777
#> 373  132200   316.22777
#> 374  145300   316.22777
#> 375  135900   316.22777
#> 376  162300   316.22777
#> 377  145300   316.22777
#> 378  141700   316.22777
#> 379  155400   316.22777
#> 380  103900   316.22777
#> 381   98800   316.22777
#> 382   96100   316.22777
#> 383   97700   316.22777
#> 384  103500   316.22777
#> 385   98300   316.22777
#> 386   88200   316.22777
#> 387  106600   316.22777
#> 388  104300   316.22777
#> 389  105700   316.22777
#> 390  104900   316.22777
#> 391   99000   316.22777
#> 392  103300   316.22777
#> 393  109900   316.22777
#> 394  104100   316.22777
#> 395  104900   316.22777
#> 396  102000   316.22777
#> 397   94000   316.22777
#> 398   66900   316.22777
#> 399   57400   316.22777
#> 400   57100   316.22777
#> 401  193600  1000.00000
#> 402  164900  1000.00000
#> 403  164800  1000.00000
#> 404  166200  1000.00000
#> 405  183000  1000.00000
#> 406  233200  1000.00000
#> 407  165200  1000.00000
#> 408  162700  1000.00000
#> 409  164100  1000.00000
#> 410  163300  1000.00000
#> 411  170300  1000.00000
#> 412  165400  1000.00000
#> 413  167100  1000.00000
#> 414  163500  1000.00000
#> 415  164200  1000.00000
#> 416  180500  1000.00000
#> 417  176200  1000.00000
#> 418  165200  1000.00000
#> 419  163600  1000.00000
#> 420  165600  1000.00000
#> 421  162500  1000.00000
#> 422  163500  1000.00000
#> 423  170500  1000.00000
#> 424  163800  1000.00000
#> 425  164100  1000.00000
#> 426  164000  1000.00000
#> 427  163800  1000.00000
#> 428  163600  1000.00000
#> 429  169300  1000.00000
#> 430  162900  1000.00000
#> 431  163400  1000.00000
#> 432  161700  1000.00000
#> 433  163500  1000.00000
#> 434  163900  1000.00000
#> 435  169400  1000.00000
#> 436  164900  1000.00000
#> 437  164400  1000.00000
#> 438  163600  1000.00000
#> 439  162700  1000.00000
#> 440  163300  1000.00000
#> 441  168200  1000.00000
#> 442  162700  1000.00000
#> 443  162300  1000.00000
#> 444  164400  1000.00000
#> 445  162300  1000.00000
#> 446  163100  1000.00000
#> 447  166100  1000.00000
#> 448  166600  1000.00000
#> 449  164800  1000.00000
#> 450  163300  1000.00000
#> 451  164100  1000.00000
#> 452  163300  1000.00000
#> 453  177600  1000.00000
#> 454  159200  1000.00000
#> 455  159300  1000.00000
#> 456  158100  1000.00000
#> 457  159000  1000.00000
#> 458  158900  1000.00000
#> 459  166400  1000.00000
#> 460  159900  1000.00000
#> 461  159800  1000.00000
#> 462  157900  1000.00000
#> 463  157800  1000.00000
#> 464  158300  1000.00000
#> 465  176100  1000.00000
#> 466  154700  1000.00000
#> 467  153500  1000.00000
#> 468  152300  1000.00000
#> 469  152600  1000.00000
#> 470  153500  1000.00000
#> 471  152600  1000.00000
#> 472  157600  1000.00000
#> 473  152300  1000.00000
#> 474  153300  1000.00000
#> 475  152900  1000.00000
#> 476  152200  1000.00000
#> 477  152800  1000.00000
#> 478  168700  1000.00000
#> 479  147900  1000.00000
#> 480  148400  1000.00000
#> 481  146700  1000.00000
#> 482  148000  1000.00000
#> 483  148500  1000.00000
#> 484  148600  1000.00000
#> 485  152200  1000.00000
#> 486  149100  1000.00000
#> 487  148900  1000.00000
#> 488  149600  1000.00000
#> 489  147700  1000.00000
#> 490  149300  1000.00000
#> 491  170400  1000.00000
#> 492  151300  1000.00000
#> 493  142800  1000.00000
#> 494  143700  1000.00000
#> 495  143500  1000.00000
#> 496  142600  1000.00000
#> 497  144000  1000.00000
#> 498  148900  1000.00000
#> 499  142300  1000.00000
#> 500  144200  1000.00000
#> 501  553900  3162.27766
#> 502  514500  3162.27766
#> 503  507300  3162.27766
#> 504  514800  3162.27766
#> 505  552900  3162.27766
#> 506  599200  3162.27766
#> 507  781300  3162.27766
#> 508  508900  3162.27766
#> 509  546600  3162.27766
#> 510  503600  3162.27766
#> 511  544100  3162.27766
#> 512  515300  3162.27766
#> 513  583900  3162.27766
#> 514  796200  3162.27766
#> 515  497300  3162.27766
#> 516  502500  3162.27766
#> 517  502500  3162.27766
#> 518  518800  3162.27766
#> 519  770500  3162.27766
#> 520  522300  3162.27766
#> 521  504000  3162.27766
#> 522  509000  3162.27766
#> 523  503900  3162.27766
#> 524  515200  3162.27766
#> 525  499700  3162.27766
#> 526  516500  3162.27766
#> 527  505700  3162.27766
#> 528  797200  3162.27766
#> 529  526200  3162.27766
#> 530  505100  3162.27766
#> 531  504900  3162.27766
#> 532  496400  3162.27766
#> 533  530000  3162.27766
#> 534  497000  3162.27766
#> 535  522300  3162.27766
#> 536  499300  3162.27766
#> 537  508900  3162.27766
#> 538  498300  3162.27766
#> 539  508400  3162.27766
#> 540  501900  3162.27766
#> 541  568200  3162.27766
#> 542  512800  3162.27766
#> 543  792100  3162.27766
#> 544  727200  3162.27766
#> 545  509500  3162.27766
#> 546  517000  3162.27766
#> 547  502700  3162.27766
#> 548  505100  3162.27766
#> 549  499600  3162.27766
#> 550  516300  3162.27766
#> 551  502800  3162.27766
#> 552  509100  3162.27766
#> 553  500000  3162.27766
#> 554  506000  3162.27766
#> 555  503800  3162.27766
#> 556  517500  3162.27766
#> 557  512100  3162.27766
#> 558  506700  3162.27766
#> 559  713000  3162.27766
#> 560  917800  3162.27766
#> 561  656600  3162.27766
#> 562  589100  3162.27766
#> 563  498800  3162.27766
#> 564  510800  3162.27766
#> 565  501900  3162.27766
#> 566  528200  3162.27766
#> 567  502100  3162.27766
#> 568  505800  3162.27766
#> 569  499400  3162.27766
#> 570  508800  3162.27766
#> 571  503600  3162.27766
#> 572 1040000  3162.27766
#> 573  509500  3162.27766
#> 574  499400  3162.27766
#> 575  503500  3162.27766
#> 576  513000  3162.27766
#> 577  500600  3162.27766
#> 578  503200  3162.27766
#> 579  500600  3162.27766
#> 580  503600  3162.27766
#> 581  505600  3162.27766
#> 582  523400  3162.27766
#> 583  508200  3162.27766
#> 584  510100  3162.27766
#> 585  476700  3162.27766
#> 586  491400  3162.27766
#> 587  451500  3162.27766
#> 588  467100  3162.27766
#> 589  483600  3162.27766
#> 590  552700  3162.27766
#> 591  440800  3162.27766
#> 592  439700  3162.27766
#> 593  462700  3162.27766
#> 594  442000  3162.27766
#> 595  446600  3162.27766
#> 596  443200  3162.27766
#> 597  444500  3162.27766
#> 598  444400  3162.27766
#> 599  452800  3162.27766
#> 600  449800  3162.27766
#> 601 1586300 10000.00000
#> 602 1585900 10000.00000
#> 603 1572000 10000.00000
#> 604 1564500 10000.00000
#> 605 1523400 10000.00000
#> 606 1578900 10000.00000
#> 607 1486000 10000.00000
#> 608 1429700 10000.00000
#> 609 1409400 10000.00000
#> 610 1385100 10000.00000
#> 611 1357400 10000.00000
#> 612 1352300 10000.00000
#> 613 1325900 10000.00000
#> 614 1318400 10000.00000
#> 615 1323800 10000.00000
#> 616 1307800 10000.00000
#> 617 1340700 10000.00000
#> 618 1317700 10000.00000
#> 619 1301700 10000.00000
#> 620 1323400 10000.00000
#> 621 1314400 10000.00000
#> 622 1302000 10000.00000
#> 623 1315000 10000.00000
#> 624 1304600 10000.00000
#> 625 1304300 10000.00000
#> 626 1301900 10000.00000
#> 627 1305500 10000.00000
#> 628 1339800 10000.00000
#> 629 1316500 10000.00000
#> 630 1326800 10000.00000
#> 631 1337800 10000.00000
#> 632 1297800 10000.00000
#> 633 1313700 10000.00000
#> 634 1325300 10000.00000
#> 635 1303800 10000.00000
#> 636 1299500 10000.00000
#> 637 1329200 10000.00000
#> 638 1314800 10000.00000
#> 639 1345900 10000.00000
#> 640 1375000 10000.00000
#> 641 1289100 10000.00000
#> 642 1853700 10000.00000
#> 643 1398200 10000.00000
#> 644 1296900 10000.00000
#> 645 1960500 10000.00000
#> 646 1320100 10000.00000
#> 647 1309700 10000.00000
#> 648 1312100 10000.00000
#> 649 1477100 10000.00000
#> 650 1311100 10000.00000
#> 651 1386000 10000.00000
#> 652 1484800 10000.00000
#> 653 1353200 10000.00000
#> 654 1338500 10000.00000
#> 655 1421200 10000.00000
#> 656 1394200 10000.00000
#> 657 1310900 10000.00000
#> 658 1317100 10000.00000
#> 659 1298200 10000.00000
#> 660 1340800 10000.00000
#> 661 1302000 10000.00000
#> 662 1367800 10000.00000
#> 663 1313000 10000.00000
#> 664 1309000 10000.00000
#> 665 1300800 10000.00000
#> 666 1312400 10000.00000
#> 667 1321200 10000.00000
#> 668 1377100 10000.00000
#> 669 1304900 10000.00000
#> 670 1294300 10000.00000
#> 671 1302600 10000.00000
#> 672 1315000 10000.00000
#> 673 1312300 10000.00000
#> 674 1312200 10000.00000
#> 675 1497200 10000.00000
#> 676 1306400 10000.00000
#> 677 1306800 10000.00000
#> 678 1314300 10000.00000
#> 679 1383000 10000.00000
#> 680 1819100 10000.00000
#> 681 1454600 10000.00000
#> 682 1388700 10000.00000
#> 683 1384500 10000.00000
#> 684 1387400 10000.00000
#> 685 1374100 10000.00000
#> 686 1379400 10000.00000
#> 687 1392300 10000.00000
#> 688 1388600 10000.00000
#> 689 1383600 10000.00000
#> 690 1404700 10000.00000
#> 691 1587000 10000.00000
#> 692 1412600 10000.00000
#> 693 1444700 10000.00000
#> 694 1412900 10000.00000
#> 695 1440100 10000.00000
#> 696 1402200 10000.00000
#> 697 1406000 10000.00000
#> 698 1433700 10000.00000
#> 699 1390900 10000.00000
#> 700 1407000 10000.00000