Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the quick sort algorithm (with sampling performed)
# against a set of increasing input data sizes:
input.sizes = 10^seq(1, 3, by = 0.5)
asymptoticTimings(sort(sample(1:100, data.sizes, replace = TRUE), method = "quick"), input.sizes)

#>     Timings Data sizes
#> 1    128600   10.00000
#> 2     49900   10.00000
#> 3     40000   10.00000
#> 4     35500   10.00000
#> 5     33600   10.00000
#> 6     32500   10.00000
#> 7     31800   10.00000
#> 8     31600   10.00000
#> 9     31600   10.00000
#> 10    31600   10.00000
#> 11    31300   10.00000
#> 12    31700   10.00000
#> 13    31600   10.00000
#> 14    31000   10.00000
#> 15    30800   10.00000
#> 16    31500   10.00000
#> 17    37800   10.00000
#> 18    32000   10.00000
#> 19    31300   10.00000
#> 20    31200   10.00000
#> 21    31300   10.00000
#> 22    31400   10.00000
#> 23    30900   10.00000
#> 24    32100   10.00000
#> 25    31400   10.00000
#> 26    32100   10.00000
#> 27    31500   10.00000
#> 28    31200   10.00000
#> 29    31500   10.00000
#> 30    31300   10.00000
#> 31    32100   10.00000
#> 32    31400   10.00000
#> 33    31200   10.00000
#> 34    30800   10.00000
#> 35    31000   10.00000
#> 36    31100   10.00000
#> 37    31300   10.00000
#> 38    31300   10.00000
#> 39    31600   10.00000
#> 40    32500   10.00000
#> 41    31400   10.00000
#> 42    32100   10.00000
#> 43    31400   10.00000
#> 44    31500   10.00000
#> 45    31200   10.00000
#> 46    31700   10.00000
#> 47    31500   10.00000
#> 48    37800   10.00000
#> 49    31800   10.00000
#> 50    31400   10.00000
#> 51    31300   10.00000
#> 52    31100   10.00000
#> 53    31900   10.00000
#> 54    31700   10.00000
#> 55    31300   10.00000
#> 56    31800   10.00000
#> 57    31700   10.00000
#> 58    31400   10.00000
#> 59    31400   10.00000
#> 60    32200   10.00000
#> 61    31300   10.00000
#> 62    32100   10.00000
#> 63    32600   10.00000
#> 64    32100   10.00000
#> 65    31500   10.00000
#> 66    31500   10.00000
#> 67    31400   10.00000
#> 68    31400   10.00000
#> 69    31500   10.00000
#> 70    32300   10.00000
#> 71    31600   10.00000
#> 72    31200   10.00000
#> 73    33400   10.00000
#> 74    35900   10.00000
#> 75    32500   10.00000
#> 76    35100   10.00000
#> 77    36500   10.00000
#> 78    38300   10.00000
#> 79    32100   10.00000
#> 80    31900   10.00000
#> 81    32600   10.00000
#> 82    31300   10.00000
#> 83    31900   10.00000
#> 84    31800   10.00000
#> 85    31400   10.00000
#> 86    31500   10.00000
#> 87    31300   10.00000
#> 88    32300   10.00000
#> 89    30900   10.00000
#> 90    36100   10.00000
#> 91    31600   10.00000
#> 92    31900   10.00000
#> 93    31900   10.00000
#> 94    32500   10.00000
#> 95    32700   10.00000
#> 96    32500   10.00000
#> 97    33900   10.00000
#> 98    32700   10.00000
#> 99    33600   10.00000
#> 100   34200   10.00000
#> 101  106300   31.62278
#> 102   46200   31.62278
#> 103   39300   31.62278
#> 104   35200   31.62278
#> 105   33500   31.62278
#> 106   32300   31.62278
#> 107   32100   31.62278
#> 108   31900   31.62278
#> 109   31500   31.62278
#> 110   37400   31.62278
#> 111   34100   31.62278
#> 112   31800   31.62278
#> 113   31800   31.62278
#> 114   31300   31.62278
#> 115   31000   31.62278
#> 116   36000   31.62278
#> 117   31200   31.62278
#> 118   30700   31.62278
#> 119   32400   31.62278
#> 120   31400   31.62278
#> 121   31100   31.62278
#> 122   30800   31.62278
#> 123   31100   31.62278
#> 124   31900   31.62278
#> 125   32100   31.62278
#> 126   31600   31.62278
#> 127   31600   31.62278
#> 128   31500   31.62278
#> 129   31500   31.62278
#> 130   31100   31.62278
#> 131   31000   31.62278
#> 132   31200   31.62278
#> 133   31500   31.62278
#> 134   31500   31.62278
#> 135   30900   31.62278
#> 136   36500   31.62278
#> 137   32100   31.62278
#> 138   30900   31.62278
#> 139   31500   31.62278
#> 140   32000   31.62278
#> 141   31200   31.62278
#> 142   31700   31.62278
#> 143   31400   31.62278
#> 144   31600   31.62278
#> 145   31100   31.62278
#> 146   31800   31.62278
#> 147   36800   31.62278
#> 148   31700   31.62278
#> 149   31500   31.62278
#> 150   31400   31.62278
#> 151   31400   31.62278
#> 152   31000   31.62278
#> 153   32500   31.62278
#> 154   32400   31.62278
#> 155   31100   31.62278
#> 156   32000   31.62278
#> 157   31700   31.62278
#> 158   31600   31.62278
#> 159   31300   31.62278
#> 160   31700   31.62278
#> 161   32400   31.62278
#> 162   37700   31.62278
#> 163   32900   31.62278
#> 164   31800   31.62278
#> 165   31700   31.62278
#> 166   32000   31.62278
#> 167   32200   31.62278
#> 168   32400   31.62278
#> 169   31600   31.62278
#> 170   32000   31.62278
#> 171   31800   31.62278
#> 172   31000   31.62278
#> 173   31500   31.62278
#> 174   32200   31.62278
#> 175   31400   31.62278
#> 176   31700   31.62278
#> 177   35200   31.62278
#> 178   34500   31.62278
#> 179   34400   31.62278
#> 180   56000   31.62278
#> 181   35300   31.62278
#> 182   32800   31.62278
#> 183   32000   31.62278
#> 184   32700   31.62278
#> 185   31700   31.62278
#> 186   36600   31.62278
#> 187   35300   31.62278
#> 188   33200   31.62278
#> 189   31300   31.62278
#> 190   31200   31.62278
#> 191   32300   31.62278
#> 192   31700   31.62278
#> 193   32000   31.62278
#> 194   32500   31.62278
#> 195   31800   31.62278
#> 196   31700   31.62278
#> 197   32200   31.62278
#> 198   31500   31.62278
#> 199   32000   31.62278
#> 200   31800   31.62278
#> 201  195700  100.00000
#> 202   48400  100.00000
#> 203   39300  100.00000
#> 204   36000  100.00000
#> 205   33700  100.00000
#> 206   32300  100.00000
#> 207   31900  100.00000
#> 208   31500  100.00000
#> 209   31700  100.00000
#> 210   31700  100.00000
#> 211   32100  100.00000
#> 212   31300  100.00000
#> 213   31400  100.00000
#> 214   31300  100.00000
#> 215   31200  100.00000
#> 216   56200  100.00000
#> 217   39400  100.00000
#> 218   33000  100.00000
#> 219   32200  100.00000
#> 220   31800  100.00000
#> 221   31400  100.00000
#> 222   31300  100.00000
#> 223   31100  100.00000
#> 224   32000  100.00000
#> 225   32300  100.00000
#> 226   32200  100.00000
#> 227   31500  100.00000
#> 228   31200  100.00000
#> 229   31000  100.00000
#> 230   31400  100.00000
#> 231   31400  100.00000
#> 232   30900  100.00000
#> 233   31400  100.00000
#> 234   31000  100.00000
#> 235   31100  100.00000
#> 236   31300  100.00000
#> 237   31000  100.00000
#> 238   31200  100.00000
#> 239   31100  100.00000
#> 240   31900  100.00000
#> 241   32300  100.00000
#> 242   31600  100.00000
#> 243   31800  100.00000
#> 244   31400  100.00000
#> 245   31200  100.00000
#> 246   38100  100.00000
#> 247   32200  100.00000
#> 248   31300  100.00000
#> 249   31300  100.00000
#> 250   31100  100.00000
#> 251   32800  100.00000
#> 252   30600  100.00000
#> 253   31000  100.00000
#> 254   31700  100.00000
#> 255   31000  100.00000
#> 256   31400  100.00000
#> 257   31800  100.00000
#> 258   31200  100.00000
#> 259   30700  100.00000
#> 260   31000  100.00000
#> 261   31600  100.00000
#> 262   30900  100.00000
#> 263   31400  100.00000
#> 264   32700  100.00000
#> 265   31600  100.00000
#> 266   31400  100.00000
#> 267   31400  100.00000
#> 268   31100  100.00000
#> 269   30900  100.00000
#> 270   33100  100.00000
#> 271   31500  100.00000
#> 272   31300  100.00000
#> 273   31300  100.00000
#> 274   31000  100.00000
#> 275   31500  100.00000
#> 276   31400  100.00000
#> 277   34200  100.00000
#> 278   35300  100.00000
#> 279   31200  100.00000
#> 280   31400  100.00000
#> 281   31700  100.00000
#> 282   31300  100.00000
#> 283   31000  100.00000
#> 284   32000  100.00000
#> 285   31600  100.00000
#> 286   32000  100.00000
#> 287   31400  100.00000
#> 288   32200  100.00000
#> 289   32400  100.00000
#> 290   31000  100.00000
#> 291   31200  100.00000
#> 292   31300  100.00000
#> 293   31600  100.00000
#> 294   31600  100.00000
#> 295   31500  100.00000
#> 296   31500  100.00000
#> 297   32300  100.00000
#> 298   31700  100.00000
#> 299   32300  100.00000
#> 300   32100  100.00000
#> 301  106900  316.22777
#> 302   45600  316.22777
#> 303   38000  316.22777
#> 304   34200  316.22777
#> 305   32900  316.22777
#> 306   32800  316.22777
#> 307   32000  316.22777
#> 308   31800  316.22777
#> 309   31300  316.22777
#> 310   31700  316.22777
#> 311   31600  316.22777
#> 312   31700  316.22777
#> 313   31900  316.22777
#> 314   31600  316.22777
#> 315   31400  316.22777
#> 316   31300  316.22777
#> 317   31100  316.22777
#> 318   31400  316.22777
#> 319   31500  316.22777
#> 320   31300  316.22777
#> 321   31300  316.22777
#> 322   31500  316.22777
#> 323   31600  316.22777
#> 324   32000  316.22777
#> 325   32000  316.22777
#> 326   32500  316.22777
#> 327   31900  316.22777
#> 328   31400  316.22777
#> 329   31800  316.22777
#> 330   31500  316.22777
#> 331   36300  316.22777
#> 332   31400  316.22777
#> 333   31600  316.22777
#> 334   31700  316.22777
#> 335   31600  316.22777
#> 336   31300  316.22777
#> 337   31400  316.22777
#> 338   32200  316.22777
#> 339   31300  316.22777
#> 340   32000  316.22777
#> 341   32200  316.22777
#> 342   32200  316.22777
#> 343   32100  316.22777
#> 344   31200  316.22777
#> 345   31100  316.22777
#> 346   31800  316.22777
#> 347   31800  316.22777
#> 348   31700  316.22777
#> 349   31900  316.22777
#> 350   31700  316.22777
#> 351   31700  316.22777
#> 352   31700  316.22777
#> 353   32200  316.22777
#> 354   32200  316.22777
#> 355   31800  316.22777
#> 356   32600  316.22777
#> 357   31900  316.22777
#> 358   31700  316.22777
#> 359   31400  316.22777
#> 360   31400  316.22777
#> 361   35200  316.22777
#> 362   34100  316.22777
#> 363   37200  316.22777
#> 364   33300  316.22777
#> 365   31700  316.22777
#> 366   32200  316.22777
#> 367   32300  316.22777
#> 368   32000  316.22777
#> 369   31600  316.22777
#> 370   32300  316.22777
#> 371   31900  316.22777
#> 372   31700  316.22777
#> 373   32200  316.22777
#> 374   32100  316.22777
#> 375   31300  316.22777
#> 376   32000  316.22777
#> 377   31900  316.22777
#> 378   32900  316.22777
#> 379   32300  316.22777
#> 380   31800  316.22777
#> 381   31800  316.22777
#> 382   31700  316.22777
#> 383   31500  316.22777
#> 384   32100  316.22777
#> 385   31400  316.22777
#> 386   32300  316.22777
#> 387   32200  316.22777
#> 388   32700  316.22777
#> 389   32800  316.22777
#> 390   31500  316.22777
#> 391   31600  316.22777
#> 392   37400  316.22777
#> 393   32200  316.22777
#> 394   31600  316.22777
#> 395   31600  316.22777
#> 396   31600  316.22777
#> 397   32300  316.22777
#> 398   31200  316.22777
#> 399   32000  316.22777
#> 400   31700  316.22777
#> 401   99100 1000.00000
#> 402   44300 1000.00000
#> 403   53800 1000.00000
#> 404   60500 1000.00000
#> 405   38600 1000.00000
#> 406   39000 1000.00000
#> 407   97800 1000.00000
#> 408   47100 1000.00000
#> 409   34900 1000.00000
#> 410   32800 1000.00000
#> 411   32100 1000.00000
#> 412   31900 1000.00000
#> 413   31700 1000.00000
#> 414   31100 1000.00000
#> 415   30900 1000.00000
#> 416   31700 1000.00000
#> 417   31300 1000.00000
#> 418   30900 1000.00000
#> 419   31000 1000.00000
#> 420   31100 1000.00000
#> 421   30700 1000.00000
#> 422   30800 1000.00000
#> 423   30700 1000.00000
#> 424   31100 1000.00000
#> 425   31700 1000.00000
#> 426   31800 1000.00000
#> 427   31400 1000.00000
#> 428   32300 1000.00000
#> 429   31100 1000.00000
#> 430   37800 1000.00000
#> 431   31600 1000.00000
#> 432   31200 1000.00000
#> 433   31100 1000.00000
#> 434   31300 1000.00000
#> 435   31000 1000.00000
#> 436   31300 1000.00000
#> 437   31300 1000.00000
#> 438   31000 1000.00000
#> 439   31000 1000.00000
#> 440   31500 1000.00000
#> 441   31800 1000.00000
#> 442   31700 1000.00000
#> 443   31300 1000.00000
#> 444   30700 1000.00000
#> 445   31200 1000.00000
#> 446   31100 1000.00000
#> 447   31000 1000.00000
#> 448   31200 1000.00000
#> 449   31000 1000.00000
#> 450   31300 1000.00000
#> 451   31400 1000.00000
#> 452   30500 1000.00000
#> 453   31100 1000.00000
#> 454   31500 1000.00000
#> 455   30900 1000.00000
#> 456   31300 1000.00000
#> 457   32100 1000.00000
#> 458   31200 1000.00000
#> 459   31100 1000.00000
#> 460   31200 1000.00000
#> 461   69600 1000.00000
#> 462   43000 1000.00000
#> 463   38100 1000.00000
#> 464   39100 1000.00000
#> 465   37400 1000.00000
#> 466   37600 1000.00000
#> 467   37600 1000.00000
#> 468   37500 1000.00000
#> 469   37400 1000.00000
#> 470   37900 1000.00000
#> 471   37600 1000.00000
#> 472   37400 1000.00000
#> 473   37700 1000.00000
#> 474   37300 1000.00000
#> 475   37000 1000.00000
#> 476   38100 1000.00000
#> 477   38500 1000.00000
#> 478   39200 1000.00000
#> 479   37800 1000.00000
#> 480   37800 1000.00000
#> 481   37700 1000.00000
#> 482   37500 1000.00000
#> 483   37900 1000.00000
#> 484   38500 1000.00000
#> 485   37200 1000.00000
#> 486   42000 1000.00000
#> 487   34700 1000.00000
#> 488   33300 1000.00000
#> 489   32800 1000.00000
#> 490   32400 1000.00000
#> 491   55500 1000.00000
#> 492   64900 1000.00000
#> 493   65500 1000.00000
#> 494   59500 1000.00000
#> 495   70700 1000.00000
#> 496   73400 1000.00000
#> 497   69200 1000.00000
#> 498   76600 1000.00000
#> 499   75300 1000.00000
#> 500   77200 1000.00000