Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the quick sort algorithm (with sampling performed)
# against a set of increasing input data sizes:
input.sizes = 10^seq(1, 3, by = 0.5)
asymptoticTimings(sort(sample(1:100, data.sizes, replace = TRUE), method = "quick"), input.sizes)

#>     Timings Data sizes
#> 1    142800   10.00000
#> 2     55100   10.00000
#> 3     42700   10.00000
#> 4     38900   10.00000
#> 5     37200   10.00000
#> 6     36300   10.00000
#> 7     35800   10.00000
#> 8     35700   10.00000
#> 9     35900   10.00000
#> 10    35100   10.00000
#> 11    35300   10.00000
#> 12    35200   10.00000
#> 13    35000   10.00000
#> 14    35400   10.00000
#> 15    35000   10.00000
#> 16    34900   10.00000
#> 17    35300   10.00000
#> 18    34600   10.00000
#> 19    35200   10.00000
#> 20    34900   10.00000
#> 21    35300   10.00000
#> 22    35100   10.00000
#> 23    34600   10.00000
#> 24    34700   10.00000
#> 25    34900   10.00000
#> 26    34700   10.00000
#> 27    41600   10.00000
#> 28    35800   10.00000
#> 29    35700   10.00000
#> 30    35500   10.00000
#> 31    35600   10.00000
#> 32    34800   10.00000
#> 33    34700   10.00000
#> 34    34300   10.00000
#> 35    34300   10.00000
#> 36    34800   10.00000
#> 37    34900   10.00000
#> 38    35400   10.00000
#> 39    34500   10.00000
#> 40    35300   10.00000
#> 41    34400   10.00000
#> 42    35700   10.00000
#> 43    34400   10.00000
#> 44    35200   10.00000
#> 45    35800   10.00000
#> 46    35500   10.00000
#> 47    35600   10.00000
#> 48    35400   10.00000
#> 49    35500   10.00000
#> 50    34900   10.00000
#> 51    35700   10.00000
#> 52    35900   10.00000
#> 53    35100   10.00000
#> 54    34700   10.00000
#> 55    42700   10.00000
#> 56    36200   10.00000
#> 57    34700   10.00000
#> 58    34800   10.00000
#> 59    35400   10.00000
#> 60    34500   10.00000
#> 61    34900   10.00000
#> 62    34200   10.00000
#> 63    35100   10.00000
#> 64    34800   10.00000
#> 65    35800   10.00000
#> 66    34000   10.00000
#> 67    35300   10.00000
#> 68    34500   10.00000
#> 69    36900   10.00000
#> 70    35500   10.00000
#> 71    35200   10.00000
#> 72    35100   10.00000
#> 73    34800   10.00000
#> 74    35100   10.00000
#> 75    36000   10.00000
#> 76    42200   10.00000
#> 77    36300   10.00000
#> 78    35500   10.00000
#> 79    39300   10.00000
#> 80    34900   10.00000
#> 81    35900   10.00000
#> 82    88800   10.00000
#> 83    84600   10.00000
#> 84    82900   10.00000
#> 85    45400   10.00000
#> 86    39800   10.00000
#> 87   135700   10.00000
#> 88    76900   10.00000
#> 89    42900   10.00000
#> 90    37900   10.00000
#> 91    38800   10.00000
#> 92    38600   10.00000
#> 93    39900   10.00000
#> 94    40300   10.00000
#> 95    40900   10.00000
#> 96    40400   10.00000
#> 97    40600   10.00000
#> 98    42500   10.00000
#> 99    44300   10.00000
#> 100   49200   10.00000
#> 101  104200   31.62278
#> 102   46500   31.62278
#> 103   38100   31.62278
#> 104   35300   31.62278
#> 105   34200   31.62278
#> 106   33200   31.62278
#> 107   32600   31.62278
#> 108   32300   31.62278
#> 109   31700   31.62278
#> 110   31800   31.62278
#> 111   32100   31.62278
#> 112   32300   31.62278
#> 113   32000   31.62278
#> 114   31900   31.62278
#> 115   31300   31.62278
#> 116   41700   31.62278
#> 117   38100   31.62278
#> 118   37100   31.62278
#> 119   38300   31.62278
#> 120   37300   31.62278
#> 121   37100   31.62278
#> 122   37700   31.62278
#> 123   37100   31.62278
#> 124   37700   31.62278
#> 125   37400   31.62278
#> 126   37100   31.62278
#> 127   37300   31.62278
#> 128   37700   31.62278
#> 129   37100   31.62278
#> 130   37400   31.62278
#> 131   33200   31.62278
#> 132   32600   31.62278
#> 133   32100   31.62278
#> 134   31400   31.62278
#> 135   31100   31.62278
#> 136   31500   31.62278
#> 137   31100   31.62278
#> 138   31200   31.62278
#> 139   31100   31.62278
#> 140   31600   31.62278
#> 141   31500   31.62278
#> 142   31600   31.62278
#> 143   31800   31.62278
#> 144   31100   31.62278
#> 145   38000   31.62278
#> 146   32500   31.62278
#> 147   31500   31.62278
#> 148   31800   31.62278
#> 149   31400   31.62278
#> 150   31900   31.62278
#> 151   31600   31.62278
#> 152   32500   31.62278
#> 153   31700   31.62278
#> 154   31900   31.62278
#> 155   31600   31.62278
#> 156   31700   31.62278
#> 157   31200   31.62278
#> 158   31300   31.62278
#> 159   31500   31.62278
#> 160   32400   31.62278
#> 161   31500   31.62278
#> 162   31500   31.62278
#> 163   31000   31.62278
#> 164   31400   31.62278
#> 165   32200   31.62278
#> 166   31800   31.62278
#> 167   31000   31.62278
#> 168   31900   31.62278
#> 169   31400   31.62278
#> 170   32700   31.62278
#> 171   32200   31.62278
#> 172   31500   31.62278
#> 173   31000   31.62278
#> 174   31200   31.62278
#> 175   33400   31.62278
#> 176   37400   31.62278
#> 177   31900   31.62278
#> 178   31100   31.62278
#> 179   31300   31.62278
#> 180   31300   31.62278
#> 181   31500   31.62278
#> 182   32000   31.62278
#> 183   31100   31.62278
#> 184   31600   31.62278
#> 185   31100   31.62278
#> 186   32000   31.62278
#> 187   31900   31.62278
#> 188   31800   31.62278
#> 189   31500   31.62278
#> 190   31100   31.62278
#> 191   31200   31.62278
#> 192   31800   31.62278
#> 193   31700   31.62278
#> 194   31700   31.62278
#> 195   31100   31.62278
#> 196   31400   31.62278
#> 197   31800   31.62278
#> 198   31100   31.62278
#> 199   31500   31.62278
#> 200   31700   31.62278
#> 201  122600  100.00000
#> 202   49500  100.00000
#> 203   40000  100.00000
#> 204   37600  100.00000
#> 205   34600  100.00000
#> 206   33900  100.00000
#> 207   33200  100.00000
#> 208   32900  100.00000
#> 209   33000  100.00000
#> 210   32900  100.00000
#> 211   33400  100.00000
#> 212   32800  100.00000
#> 213   32900  100.00000
#> 214   32700  100.00000
#> 215   32500  100.00000
#> 216   32400  100.00000
#> 217   32700  100.00000
#> 218   32800  100.00000
#> 219   32700  100.00000
#> 220   32600  100.00000
#> 221   32200  100.00000
#> 222   33200  100.00000
#> 223   32400  100.00000
#> 224   32500  100.00000
#> 225   33300  100.00000
#> 226   68300  100.00000
#> 227   39600  100.00000
#> 228   38800  100.00000
#> 229   38500  100.00000
#> 230   39300  100.00000
#> 231   38800  100.00000
#> 232   38700  100.00000
#> 233   38500  100.00000
#> 234   38600  100.00000
#> 235   38200  100.00000
#> 236   38300  100.00000
#> 237   39000  100.00000
#> 238   38500  100.00000
#> 239   38700  100.00000
#> 240   38700  100.00000
#> 241   38600  100.00000
#> 242   38900  100.00000
#> 243   38700  100.00000
#> 244   38500  100.00000
#> 245   39200  100.00000
#> 246   39300  100.00000
#> 247   38600  100.00000
#> 248   39100  100.00000
#> 249   38300  100.00000
#> 250  104800  100.00000
#> 251   97400  100.00000
#> 252   54700  100.00000
#> 253   45100  100.00000
#> 254   42100  100.00000
#> 255   40600  100.00000
#> 256   44700  100.00000
#> 257   39600  100.00000
#> 258   38900  100.00000
#> 259   38700  100.00000
#> 260   39600  100.00000
#> 261   38400  100.00000
#> 262   38400  100.00000
#> 263   38600  100.00000
#> 264   38500  100.00000
#> 265   38800  100.00000
#> 266   38600  100.00000
#> 267   38300  100.00000
#> 268   43300  100.00000
#> 269   38200  100.00000
#> 270   40000  100.00000
#> 271   42600  100.00000
#> 272   80600  100.00000
#> 273  234900  100.00000
#> 274  140900  100.00000
#> 275   40900  100.00000
#> 276   36700  100.00000
#> 277   38100  100.00000
#> 278   33700  100.00000
#> 279   33300  100.00000
#> 280   33100  100.00000
#> 281   33900  100.00000
#> 282   32900  100.00000
#> 283   32800  100.00000
#> 284   32800  100.00000
#> 285   32600  100.00000
#> 286   33500  100.00000
#> 287   34700  100.00000
#> 288   33000  100.00000
#> 289   32600  100.00000
#> 290   41300  100.00000
#> 291   34400  100.00000
#> 292   33900  100.00000
#> 293   33600  100.00000
#> 294   33400  100.00000
#> 295   32500  100.00000
#> 296   32700  100.00000
#> 297   33200  100.00000
#> 298   32600  100.00000
#> 299   32500  100.00000
#> 300   34200  100.00000
#> 301  102800  316.22777
#> 302   44900  316.22777
#> 303   37800  316.22777
#> 304   34200  316.22777
#> 305   32700  316.22777
#> 306   32600  316.22777
#> 307   31700  316.22777
#> 308   31500  316.22777
#> 309   31100  316.22777
#> 310   31600  316.22777
#> 311   31300  316.22777
#> 312   31500  316.22777
#> 313   35300  316.22777
#> 314   33900  316.22777
#> 315   31400  316.22777
#> 316   30800  316.22777
#> 317   31000  316.22777
#> 318   31200  316.22777
#> 319   30900  316.22777
#> 320   30800  316.22777
#> 321   30800  316.22777
#> 322   31900  316.22777
#> 323   31600  316.22777
#> 324   31700  316.22777
#> 325   31500  316.22777
#> 326   31300  316.22777
#> 327   31200  316.22777
#> 328   31900  316.22777
#> 329   31400  316.22777
#> 330   31400  316.22777
#> 331   31300  316.22777
#> 332   31400  316.22777
#> 333   31600  316.22777
#> 334   31100  316.22777
#> 335   30800  316.22777
#> 336   31300  316.22777
#> 337   30900  316.22777
#> 338   31700  316.22777
#> 339   31200  316.22777
#> 340   31100  316.22777
#> 341   31400  316.22777
#> 342   31800  316.22777
#> 343   31300  316.22777
#> 344   31000  316.22777
#> 345   37200  316.22777
#> 346   32400  316.22777
#> 347   31300  316.22777
#> 348   32000  316.22777
#> 349   31100  316.22777
#> 350   31300  316.22777
#> 351   31700  316.22777
#> 352   32400  316.22777
#> 353   31500  316.22777
#> 354   31600  316.22777
#> 355   31300  316.22777
#> 356   31600  316.22777
#> 357   31200  316.22777
#> 358   30800  316.22777
#> 359   31200  316.22777
#> 360   32100  316.22777
#> 361   31900  316.22777
#> 362   31400  316.22777
#> 363   31200  316.22777
#> 364   31300  316.22777
#> 365   31800  316.22777
#> 366   31200  316.22777
#> 367   31000  316.22777
#> 368   31000  316.22777
#> 369   31100  316.22777
#> 370   32300  316.22777
#> 371   31900  316.22777
#> 372   31900  316.22777
#> 373   31800  316.22777
#> 374   31000  316.22777
#> 375   32700  316.22777
#> 376   38800  316.22777
#> 377   31700  316.22777
#> 378   31200  316.22777
#> 379   31000  316.22777
#> 380   31200  316.22777
#> 381   31900  316.22777
#> 382   31600  316.22777
#> 383   31700  316.22777
#> 384   31400  316.22777
#> 385   31700  316.22777
#> 386   31300  316.22777
#> 387   31800  316.22777
#> 388   31700  316.22777
#> 389   32000  316.22777
#> 390   31200  316.22777
#> 391   31600  316.22777
#> 392   31900  316.22777
#> 393   32200  316.22777
#> 394   31400  316.22777
#> 395   31800  316.22777
#> 396   31000  316.22777
#> 397   31400  316.22777
#> 398   30700  316.22777
#> 399   31200  316.22777
#> 400   32100  316.22777
#> 401  119700 1000.00000
#> 402   49600 1000.00000
#> 403   42900 1000.00000
#> 404   39900 1000.00000
#> 405   38700 1000.00000
#> 406   38100 1000.00000
#> 407   37500 1000.00000
#> 408   37300 1000.00000
#> 409   37000 1000.00000
#> 410   37000 1000.00000
#> 411   37500 1000.00000
#> 412   37600 1000.00000
#> 413   37300 1000.00000
#> 414   37500 1000.00000
#> 415   37300 1000.00000
#> 416   36900 1000.00000
#> 417   36900 1000.00000
#> 418   36900 1000.00000
#> 419   37300 1000.00000
#> 420   37000 1000.00000
#> 421   36800 1000.00000
#> 422   40600 1000.00000
#> 423   32500 1000.00000
#> 424   32000 1000.00000
#> 425   37200 1000.00000
#> 426   38100 1000.00000
#> 427   37500 1000.00000
#> 428   38600 1000.00000
#> 429   38000 1000.00000
#> 430   38000 1000.00000
#> 431   37500 1000.00000
#> 432   38100 1000.00000
#> 433   38300 1000.00000
#> 434   37900 1000.00000
#> 435   37000 1000.00000
#> 436   37200 1000.00000
#> 437   37200 1000.00000
#> 438   37900 1000.00000
#> 439   37400 1000.00000
#> 440   37400 1000.00000
#> 441   37700 1000.00000
#> 442   39200 1000.00000
#> 443   39200 1000.00000
#> 444   37100 1000.00000
#> 445   38800 1000.00000
#> 446   39800 1000.00000
#> 447   38000 1000.00000
#> 448   69500 1000.00000
#> 449   48300 1000.00000
#> 450   35900 1000.00000
#> 451   33800 1000.00000
#> 452   33800 1000.00000
#> 453   32400 1000.00000
#> 454   32800 1000.00000
#> 455   31600 1000.00000
#> 456   32300 1000.00000
#> 457   31700 1000.00000
#> 458   31300 1000.00000
#> 459   31600 1000.00000
#> 460   32600 1000.00000
#> 461   32700 1000.00000
#> 462   32800 1000.00000
#> 463   32800 1000.00000
#> 464   32300 1000.00000
#> 465   32500 1000.00000
#> 466   32000 1000.00000
#> 467   32000 1000.00000
#> 468   31700 1000.00000
#> 469   31900 1000.00000
#> 470   33300 1000.00000
#> 471   32600 1000.00000
#> 472   32600 1000.00000
#> 473   31900 1000.00000
#> 474   32300 1000.00000
#> 475   33500 1000.00000
#> 476   33800 1000.00000
#> 477   41500 1000.00000
#> 478   32700 1000.00000
#> 479   32100 1000.00000
#> 480   32000 1000.00000
#> 481   32000 1000.00000
#> 482   32100 1000.00000
#> 483   54900 1000.00000
#> 484   47800 1000.00000
#> 485   32400 1000.00000
#> 486   31800 1000.00000
#> 487   32800 1000.00000
#> 488   31500 1000.00000
#> 489   31100 1000.00000
#> 490   31100 1000.00000
#> 491   31000 1000.00000
#> 492   31200 1000.00000
#> 493   32200 1000.00000
#> 494   32200 1000.00000
#> 495   31400 1000.00000
#> 496   31100 1000.00000
#> 497   31700 1000.00000
#> 498   30400 1000.00000
#> 499   31200 1000.00000
#> 500   32000 1000.00000