Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the substring function against a set of input data sizes:
input.sizes = 10^seq(1, 4, by = 0.5)
asymptoticTimings(substring(paste(rep("A", N), collapse = ""), 1:N, 1:N), input.sizes)

#>     Timings  Data sizes
#> 1     54900    10.00000
#> 2     13700    10.00000
#> 3     10000    10.00000
#> 4      8700    10.00000
#> 5      8500    10.00000
#> 6      8700    10.00000
#> 7      8700    10.00000
#> 8      8500    10.00000
#> 9      8700    10.00000
#> 10     8900    10.00000
#> 11     8500    10.00000
#> 12     8500    10.00000
#> 13    22100    10.00000
#> 14     8200    10.00000
#> 15     8300    10.00000
#> 16     8100    10.00000
#> 17     8300    10.00000
#> 18     8200    10.00000
#> 19     8200    10.00000
#> 20     8300    10.00000
#> 21     8200    10.00000
#> 22     8200    10.00000
#> 23     8200    10.00000
#> 24     8400    10.00000
#> 25     8000    10.00000
#> 26     8400    10.00000
#> 27     8100    10.00000
#> 28     8400    10.00000
#> 29    12800    10.00000
#> 30     8400    10.00000
#> 31     8400    10.00000
#> 32     8300    10.00000
#> 33     8300    10.00000
#> 34     8400    10.00000
#> 35     8200    10.00000
#> 36     8200    10.00000
#> 37     8600    10.00000
#> 38     8000    10.00000
#> 39     8800    10.00000
#> 40     8300    10.00000
#> 41     8200    10.00000
#> 42     8300    10.00000
#> 43     8500    10.00000
#> 44     8400    10.00000
#> 45     8400    10.00000
#> 46     8300    10.00000
#> 47     8300    10.00000
#> 48     8200    10.00000
#> 49     8300    10.00000
#> 50     8400    10.00000
#> 51     8200    10.00000
#> 52     8300    10.00000
#> 53     8300    10.00000
#> 54     8100    10.00000
#> 55     8700    10.00000
#> 56     8300    10.00000
#> 57     8100    10.00000
#> 58     8400    10.00000
#> 59     8200    10.00000
#> 60     8400    10.00000
#> 61     8200    10.00000
#> 62     8100    10.00000
#> 63     8200    10.00000
#> 64     8100    10.00000
#> 65     8400    10.00000
#> 66     8200    10.00000
#> 67     8300    10.00000
#> 68     8600    10.00000
#> 69     8200    10.00000
#> 70     8300    10.00000
#> 71     8300    10.00000
#> 72     8200    10.00000
#> 73     8100    10.00000
#> 74     8500    10.00000
#> 75     8100    10.00000
#> 76     8200    10.00000
#> 77     8300    10.00000
#> 78     8300    10.00000
#> 79     8200    10.00000
#> 80     8400    10.00000
#> 81     8300    10.00000
#> 82     8400    10.00000
#> 83     8100    10.00000
#> 84     8100    10.00000
#> 85     8300    10.00000
#> 86     8200    10.00000
#> 87     8100    10.00000
#> 88     8100    10.00000
#> 89     8700    10.00000
#> 90     8100    10.00000
#> 91     8500    10.00000
#> 92     8200    10.00000
#> 93     8700    10.00000
#> 94     8700    10.00000
#> 95     8200    10.00000
#> 96     8700    10.00000
#> 97     9100    10.00000
#> 98     8300    10.00000
#> 99     8800    10.00000
#> 100    8500    10.00000
#> 101   44200    31.62278
#> 102   20300    31.62278
#> 103   17500    31.62278
#> 104   17000    31.62278
#> 105   16600    31.62278
#> 106   16900    31.62278
#> 107   16600    31.62278
#> 108   25400    31.62278
#> 109   17900    31.62278
#> 110   16500    31.62278
#> 111   16400    31.62278
#> 112   16800    31.62278
#> 113   16300    31.62278
#> 114   16500    31.62278
#> 115   16200    31.62278
#> 116   16700    31.62278
#> 117   16300    31.62278
#> 118   16300    31.62278
#> 119   16600    31.62278
#> 120   16500    31.62278
#> 121   16600    31.62278
#> 122   16300    31.62278
#> 123   16400    31.62278
#> 124   16900    31.62278
#> 125   16900    31.62278
#> 126   16900    31.62278
#> 127   16600    31.62278
#> 128   16900    31.62278
#> 129   16700    31.62278
#> 130   16700    31.62278
#> 131   16700    31.62278
#> 132   16800    31.62278
#> 133   16500    31.62278
#> 134   16600    31.62278
#> 135   16600    31.62278
#> 136   16200    31.62278
#> 137   16500    31.62278
#> 138   16600    31.62278
#> 139   16800    31.62278
#> 140   16600    31.62278
#> 141   16600    31.62278
#> 142   17800    31.62278
#> 143   17000    31.62278
#> 144   16400    31.62278
#> 145   16800    31.62278
#> 146   16600    31.62278
#> 147   16400    31.62278
#> 148   16900    31.62278
#> 149   16500    31.62278
#> 150   16600    31.62278
#> 151   17100    31.62278
#> 152   16600    31.62278
#> 153   16900    31.62278
#> 154   16900    31.62278
#> 155   16900    31.62278
#> 156   16700    31.62278
#> 157   16400    31.62278
#> 158   16800    31.62278
#> 159   16400    31.62278
#> 160   16800    31.62278
#> 161   16600    31.62278
#> 162   16400    31.62278
#> 163   16700    31.62278
#> 164   16700    31.62278
#> 165   16400    31.62278
#> 166   21400    31.62278
#> 167   17300    31.62278
#> 168   16700    31.62278
#> 169   16400    31.62278
#> 170   12600    31.62278
#> 171   12300    31.62278
#> 172   12000    31.62278
#> 173   12100    31.62278
#> 174   12100    31.62278
#> 175   11800    31.62278
#> 176   12200    31.62278
#> 177   12100    31.62278
#> 178   11800    31.62278
#> 179   12100    31.62278
#> 180   12000    31.62278
#> 181   11700    31.62278
#> 182   12500    31.62278
#> 183   12000    31.62278
#> 184   12200    31.62278
#> 185   12200    31.62278
#> 186   12300    31.62278
#> 187   12100    31.62278
#> 188   12400    31.62278
#> 189   12000    31.62278
#> 190   11800    31.62278
#> 191   12200    31.62278
#> 192   11700    31.62278
#> 193   11900    31.62278
#> 194   11800    31.62278
#> 195   12000    31.62278
#> 196   12500    31.62278
#> 197   12600    31.62278
#> 198   12300    31.62278
#> 199   12300    31.62278
#> 200   12100    31.62278
#> 201   61200   100.00000
#> 202   27600   100.00000
#> 203   24400   100.00000
#> 204   23400   100.00000
#> 205   23200   100.00000
#> 206   31100   100.00000
#> 207   23600   100.00000
#> 208   24200   100.00000
#> 209   24300   100.00000
#> 210   23600   100.00000
#> 211   35600   100.00000
#> 212   60700   100.00000
#> 213   50900   100.00000
#> 214   58300   100.00000
#> 215   31300   100.00000
#> 216   24300   100.00000
#> 217   23400   100.00000
#> 218   22900   100.00000
#> 219   23500   100.00000
#> 220   23500   100.00000
#> 221   23000   100.00000
#> 222   22900   100.00000
#> 223   23000   100.00000
#> 224   22700   100.00000
#> 225   23300   100.00000
#> 226   23000   100.00000
#> 227   23000   100.00000
#> 228   23200   100.00000
#> 229   23700   100.00000
#> 230   23700   100.00000
#> 231   24000   100.00000
#> 232   23500   100.00000
#> 233   22600   100.00000
#> 234   23700   100.00000
#> 235   23500   100.00000
#> 236   23900   100.00000
#> 237   24100   100.00000
#> 238   23100   100.00000
#> 239   23300   100.00000
#> 240   25000   100.00000
#> 241   23500   100.00000
#> 242   28400   100.00000
#> 243   24800   100.00000
#> 244   23100   100.00000
#> 245   23000   100.00000
#> 246   23700   100.00000
#> 247   23700   100.00000
#> 248   23500   100.00000
#> 249   23100   100.00000
#> 250   23100   100.00000
#> 251   24100   100.00000
#> 252   24600   100.00000
#> 253   24000   100.00000
#> 254   23700   100.00000
#> 255   24000   100.00000
#> 256   23300   100.00000
#> 257   23600   100.00000
#> 258   24100   100.00000
#> 259   22900   100.00000
#> 260   23300   100.00000
#> 261   24100   100.00000
#> 262   23400   100.00000
#> 263   43800   100.00000
#> 264   55300   100.00000
#> 265   50900   100.00000
#> 266   50600   100.00000
#> 267   49400   100.00000
#> 268   48600   100.00000
#> 269   49400   100.00000
#> 270   57500   100.00000
#> 271   58000   100.00000
#> 272   91900   100.00000
#> 273   49000   100.00000
#> 274   36200   100.00000
#> 275   33600   100.00000
#> 276   34000   100.00000
#> 277   34100   100.00000
#> 278   33500   100.00000
#> 279   33900   100.00000
#> 280   34000   100.00000
#> 281   34800   100.00000
#> 282   34000   100.00000
#> 283   34400   100.00000
#> 284   33800   100.00000
#> 285   33800   100.00000
#> 286   33600   100.00000
#> 287   37600   100.00000
#> 288   39800   100.00000
#> 289   44400   100.00000
#> 290   49400   100.00000
#> 291   56400   100.00000
#> 292   51400   100.00000
#> 293   43600   100.00000
#> 294   48700   100.00000
#> 295   41400   100.00000
#> 296   39000   100.00000
#> 297   27400   100.00000
#> 298   23800   100.00000
#> 299   28100   100.00000
#> 300   23600   100.00000
#> 301  116800   316.22777
#> 302   91200   316.22777
#> 303   92300   316.22777
#> 304   89800   316.22777
#> 305   87400   316.22777
#> 306   86800   316.22777
#> 307   86500   316.22777
#> 308   86100   316.22777
#> 309   88900   316.22777
#> 310   87100   316.22777
#> 311   86600   316.22777
#> 312   88400   316.22777
#> 313   87300   316.22777
#> 314   85300   316.22777
#> 315   92500   316.22777
#> 316   87100   316.22777
#> 317   86100   316.22777
#> 318   85600   316.22777
#> 319   87200   316.22777
#> 320   86600   316.22777
#> 321   88300   316.22777
#> 322   86600   316.22777
#> 323   85800   316.22777
#> 324   89900   316.22777
#> 325   87500   316.22777
#> 326  110200   316.22777
#> 327  115600   316.22777
#> 328   89200   316.22777
#> 329   86000   316.22777
#> 330   85300   316.22777
#> 331   86500   316.22777
#> 332   86100   316.22777
#> 333   84700   316.22777
#> 334   85600   316.22777
#> 335   85700   316.22777
#> 336   86000   316.22777
#> 337  111900   316.22777
#> 338   93300   316.22777
#> 339   87900   316.22777
#> 340   86200   316.22777
#> 341   86600   316.22777
#> 342   98100   316.22777
#> 343   84100   316.22777
#> 344   83400   316.22777
#> 345   83600   316.22777
#> 346   83400   316.22777
#> 347   83300   316.22777
#> 348   86100   316.22777
#> 349   84600   316.22777
#> 350   55300   316.22777
#> 351   55100   316.22777
#> 352   54800   316.22777
#> 353   54600   316.22777
#> 354   54400   316.22777
#> 355   55100   316.22777
#> 356   54600   316.22777
#> 357   54400   316.22777
#> 358   54900   316.22777
#> 359   54100   316.22777
#> 360   55500   316.22777
#> 361   55400   316.22777
#> 362   55900   316.22777
#> 363   54800   316.22777
#> 364   55300   316.22777
#> 365   59700   316.22777
#> 366   86000   316.22777
#> 367   81900   316.22777
#> 368   81800   316.22777
#> 369   81500   316.22777
#> 370   82400   316.22777
#> 371   94600   316.22777
#> 372   79100   316.22777
#> 373   79700   316.22777
#> 374   79900   316.22777
#> 375   78100   316.22777
#> 376   79700   316.22777
#> 377   86000   316.22777
#> 378   85400   316.22777
#> 379   79900   316.22777
#> 380   79400   316.22777
#> 381   79500   316.22777
#> 382   79500   316.22777
#> 383   79700   316.22777
#> 384   79700   316.22777
#> 385   80200   316.22777
#> 386   80000   316.22777
#> 387   79500   316.22777
#> 388   80000   316.22777
#> 389   80200   316.22777
#> 390  105200   316.22777
#> 391   84600   316.22777
#> 392   82100   316.22777
#> 393   80800   316.22777
#> 394   81400   316.22777
#> 395   80900   316.22777
#> 396   81200   316.22777
#> 397   82500   316.22777
#> 398   81600   316.22777
#> 399   80900   316.22777
#> 400   81200   316.22777
#> 401  420300  1000.00000
#> 402  187300  1000.00000
#> 403  166300  1000.00000
#> 404  170900  1000.00000
#> 405  164300  1000.00000
#> 406  162600  1000.00000
#> 407  164500  1000.00000
#> 408  163900  1000.00000
#> 409  165800  1000.00000
#> 410  175100  1000.00000
#> 411  290200  1000.00000
#> 412  316000  1000.00000
#> 413  742300  1000.00000
#> 414  246900  1000.00000
#> 415  170800  1000.00000
#> 416  173000  1000.00000
#> 417  163800  1000.00000
#> 418  164200  1000.00000
#> 419  165600  1000.00000
#> 420  162500  1000.00000
#> 421  163000  1000.00000
#> 422  166200  1000.00000
#> 423  167800  1000.00000
#> 424  163600  1000.00000
#> 425  164800  1000.00000
#> 426  165000  1000.00000
#> 427  160600  1000.00000
#> 428  188800  1000.00000
#> 429  173900  1000.00000
#> 430  163900  1000.00000
#> 431  161500  1000.00000
#> 432  163700  1000.00000
#> 433  162300  1000.00000
#> 434  170100  1000.00000
#> 435  164100  1000.00000
#> 436  167000  1000.00000
#> 437  165200  1000.00000
#> 438  175300  1000.00000
#> 439  159100  1000.00000
#> 440  173200  1000.00000
#> 441  159600  1000.00000
#> 442  156900  1000.00000
#> 443  157900  1000.00000
#> 444  172800  1000.00000
#> 445  161800  1000.00000
#> 446  169200  1000.00000
#> 447  292400  1000.00000
#> 448  176300  1000.00000
#> 449  176700  1000.00000
#> 450  157500  1000.00000
#> 451  168600  1000.00000
#> 452  163100  1000.00000
#> 453  157900  1000.00000
#> 454  158900  1000.00000
#> 455  157700  1000.00000
#> 456  157600  1000.00000
#> 457  158900  1000.00000
#> 458  166600  1000.00000
#> 459  159800  1000.00000
#> 460  158800  1000.00000
#> 461  160000  1000.00000
#> 462  157700  1000.00000
#> 463  161900  1000.00000
#> 464  170900  1000.00000
#> 465  324800  1000.00000
#> 466  299100  1000.00000
#> 467  208500  1000.00000
#> 468  216300  1000.00000
#> 469  166200  1000.00000
#> 470  163000  1000.00000
#> 471  162100  1000.00000
#> 472  163800  1000.00000
#> 473  166900  1000.00000
#> 474  165100  1000.00000
#> 475  161700  1000.00000
#> 476  161700  1000.00000
#> 477  162400  1000.00000
#> 478  171300  1000.00000
#> 479  159300  1000.00000
#> 480  163000  1000.00000
#> 481  156400  1000.00000
#> 482  158000  1000.00000
#> 483  157300  1000.00000
#> 484  167800  1000.00000
#> 485  152600  1000.00000
#> 486  158500  1000.00000
#> 487  152600  1000.00000
#> 488  153300  1000.00000
#> 489  152300  1000.00000
#> 490  162500  1000.00000
#> 491  148200  1000.00000
#> 492  166700  1000.00000
#> 493  167200  1000.00000
#> 494  150900  1000.00000
#> 495  147900  1000.00000
#> 496  147100  1000.00000
#> 497  157900  1000.00000
#> 498  144100  1000.00000
#> 499  217600  1000.00000
#> 500  221300  1000.00000
#> 501 1083000  3162.27766
#> 502  755400  3162.27766
#> 503  500700  3162.27766
#> 504  520300  3162.27766
#> 505  499600  3162.27766
#> 506  542400  3162.27766
#> 507  769700  3162.27766
#> 508  521900  3162.27766
#> 509  499600  3162.27766
#> 510  504800  3162.27766
#> 511  515100  3162.27766
#> 512  488800  3162.27766
#> 513  478000  3162.27766
#> 514  729000  3162.27766
#> 515  778000  3162.27766
#> 516  695000  3162.27766
#> 517  472700  3162.27766
#> 518  482700  3162.27766
#> 519  456200  3162.27766
#> 520  456800  3162.27766
#> 521  454900  3162.27766
#> 522  460300  3162.27766
#> 523  457500  3162.27766
#> 524  441000  3162.27766
#> 525  445300  3162.27766
#> 526  446300  3162.27766
#> 527  454300  3162.27766
#> 528  428700  3162.27766
#> 529  433700  3162.27766
#> 530  433100  3162.27766
#> 531  429700  3162.27766
#> 532  433300  3162.27766
#> 533  430000  3162.27766
#> 534  455100  3162.27766
#> 535  417200  3162.27766
#> 536  423000  3162.27766
#> 537  419200  3162.27766
#> 538  414500  3162.27766
#> 539  417000  3162.27766
#> 540  415500  3162.27766
#> 541  425200  3162.27766
#> 542  416800  3162.27766
#> 543  418700  3162.27766
#> 544  420500  3162.27766
#> 545  417500  3162.27766
#> 546  420000  3162.27766
#> 547  416700  3162.27766
#> 548  437800  3162.27766
#> 549  425000  3162.27766
#> 550  422600  3162.27766
#> 551  425800  3162.27766
#> 552  418100  3162.27766
#> 553  419600  3162.27766
#> 554  418300  3162.27766
#> 555  420900  3162.27766
#> 556  419400  3162.27766
#> 557  414700  3162.27766
#> 558  419900  3162.27766
#> 559  413600  3162.27766
#> 560  419300  3162.27766
#> 561  420400  3162.27766
#> 562  423600  3162.27766
#> 563  416600  3162.27766
#> 564  413000  3162.27766
#> 565  426600  3162.27766
#> 566  415400  3162.27766
#> 567  422500  3162.27766
#> 568  418100  3162.27766
#> 569  415500  3162.27766
#> 570  431000  3162.27766
#> 571  417700  3162.27766
#> 572  418200  3162.27766
#> 573  415600  3162.27766
#> 574  423700  3162.27766
#> 575  419600  3162.27766
#> 576  418200  3162.27766
#> 577  420400  3162.27766
#> 578  416000  3162.27766
#> 579  420700  3162.27766
#> 580  416900  3162.27766
#> 581  441500  3162.27766
#> 582  423600  3162.27766
#> 583  419500  3162.27766
#> 584  419700  3162.27766
#> 585  415800  3162.27766
#> 586  421400  3162.27766
#> 587  420000  3162.27766
#> 588  424100  3162.27766
#> 589  413600  3162.27766
#> 590  418000  3162.27766
#> 591  423500  3162.27766
#> 592  416800  3162.27766
#> 593  417800  3162.27766
#> 594  417500  3162.27766
#> 595  418800  3162.27766
#> 596  421600  3162.27766
#> 597  419500  3162.27766
#> 598  424600  3162.27766
#> 599  417000  3162.27766
#> 600  421800  3162.27766
#> 601 2543300 10000.00000
#> 602 2477500 10000.00000
#> 603 2411600 10000.00000
#> 604 2439100 10000.00000
#> 605 2364700 10000.00000
#> 606 2294900 10000.00000
#> 607 2226700 10000.00000
#> 608 2016500 10000.00000
#> 609 1385600 10000.00000
#> 610 1319000 10000.00000
#> 611 1299400 10000.00000
#> 612 1318100 10000.00000
#> 613 1304000 10000.00000
#> 614 1334500 10000.00000
#> 615 1326200 10000.00000
#> 616 1304100 10000.00000
#> 617 1302400 10000.00000
#> 618 1325300 10000.00000
#> 619 1395400 10000.00000
#> 620 1361900 10000.00000
#> 621 1353700 10000.00000
#> 622 1311800 10000.00000
#> 623 1304900 10000.00000
#> 624 1341200 10000.00000
#> 625 1301000 10000.00000
#> 626 1306800 10000.00000
#> 627 1306100 10000.00000
#> 628 1309900 10000.00000
#> 629 1312200 10000.00000
#> 630 1305300 10000.00000
#> 631 1315600 10000.00000
#> 632 1305300 10000.00000
#> 633 1335900 10000.00000
#> 634 1318700 10000.00000
#> 635 1322400 10000.00000
#> 636 1342500 10000.00000
#> 637 1322300 10000.00000
#> 638 1302500 10000.00000
#> 639 1313300 10000.00000
#> 640 1311700 10000.00000
#> 641 1305700 10000.00000
#> 642 1313900 10000.00000
#> 643 1337300 10000.00000
#> 644 1379100 10000.00000
#> 645 1345900 10000.00000
#> 646 1697100 10000.00000
#> 647 1311200 10000.00000
#> 648 1325700 10000.00000
#> 649 1294400 10000.00000
#> 650 1307700 10000.00000
#> 651 1685600 10000.00000
#> 652 1298700 10000.00000
#> 653 1312300 10000.00000
#> 654 1303100 10000.00000
#> 655 1301800 10000.00000
#> 656 1312100 10000.00000
#> 657 1305900 10000.00000
#> 658 1749200 10000.00000
#> 659 1587700 10000.00000
#> 660 1751200 10000.00000
#> 661 1811900 10000.00000
#> 662 1338900 10000.00000
#> 663 1331400 10000.00000
#> 664 1323800 10000.00000
#> 665 1293000 10000.00000
#> 666 1313300 10000.00000
#> 667 1317300 10000.00000
#> 668 1304700 10000.00000
#> 669 1864300 10000.00000
#> 670 1304500 10000.00000
#> 671 1306500 10000.00000
#> 672 1293900 10000.00000
#> 673 1300800 10000.00000
#> 674 1306100 10000.00000
#> 675 1307100 10000.00000
#> 676 1299600 10000.00000
#> 677 1315900 10000.00000
#> 678 1311500 10000.00000
#> 679 1324700 10000.00000
#> 680 1308700 10000.00000
#> 681 1464700 10000.00000
#> 682 1398300 10000.00000
#> 683 1391400 10000.00000
#> 684 1392200 10000.00000
#> 685 1417700 10000.00000
#> 686 1407400 10000.00000
#> 687 1408800 10000.00000
#> 688 1417700 10000.00000
#> 689 1402900 10000.00000
#> 690 1873200 10000.00000
#> 691 1452400 10000.00000
#> 692 1405900 10000.00000
#> 693 1402600 10000.00000
#> 694 1802100 10000.00000
#> 695 1468100 10000.00000
#> 696 1919000 10000.00000
#> 697 1396100 10000.00000
#> 698 1412600 10000.00000
#> 699 1414100 10000.00000
#> 700 1408700 10000.00000