Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the substring function against a set of input data sizes:
input.sizes = 10^seq(1, 4, by = 0.5)
asymptoticTimings(substring(paste(rep("A", N), collapse = ""), 1:N, 1:N), input.sizes)

#>     Timings  Data sizes
#> 1     53900    10.00000
#> 2     14600    10.00000
#> 3     10800    10.00000
#> 4      9400    10.00000
#> 5      9300    10.00000
#> 6      9100    10.00000
#> 7      9300    10.00000
#> 8      9100    10.00000
#> 9      9300    10.00000
#> 10     9300    10.00000
#> 11     9400    10.00000
#> 12     9300    10.00000
#> 13     9200    10.00000
#> 14     9200    10.00000
#> 15     9300    10.00000
#> 16     9300    10.00000
#> 17     9100    10.00000
#> 18     9300    10.00000
#> 19     9100    10.00000
#> 20     9300    10.00000
#> 21     9600    10.00000
#> 22     9300    10.00000
#> 23     9200    10.00000
#> 24    10600    10.00000
#> 25     9100    10.00000
#> 26    10900    10.00000
#> 27     9200    10.00000
#> 28    10900    10.00000
#> 29     9200    10.00000
#> 30     9400    10.00000
#> 31     9100    10.00000
#> 32     9300    10.00000
#> 33     9200    10.00000
#> 34     9300    10.00000
#> 35     9200    10.00000
#> 36     9100    10.00000
#> 37     9400    10.00000
#> 38     9200    10.00000
#> 39     9200    10.00000
#> 40     9100    10.00000
#> 41     9200    10.00000
#> 42     9300    10.00000
#> 43     9100    10.00000
#> 44     9200    10.00000
#> 45     9800    10.00000
#> 46     9100    10.00000
#> 47     9100    10.00000
#> 48     9400    10.00000
#> 49    12600    10.00000
#> 50    13600    10.00000
#> 51     9800    10.00000
#> 52     9200    10.00000
#> 53     9400    10.00000
#> 54     9100    10.00000
#> 55     9400    10.00000
#> 56     9200    10.00000
#> 57     9100    10.00000
#> 58     9200    10.00000
#> 59     9100    10.00000
#> 60     9400    10.00000
#> 61     9400    10.00000
#> 62     9000    10.00000
#> 63     9200    10.00000
#> 64     8900    10.00000
#> 65     9000    10.00000
#> 66     9000    10.00000
#> 67     9000    10.00000
#> 68     9100    10.00000
#> 69     9100    10.00000
#> 70     9200    10.00000
#> 71     9100    10.00000
#> 72     9200    10.00000
#> 73     9100    10.00000
#> 74     9200    10.00000
#> 75     9100    10.00000
#> 76     9300    10.00000
#> 77     9200    10.00000
#> 78     9300    10.00000
#> 79     9400    10.00000
#> 80     9300    10.00000
#> 81     9100    10.00000
#> 82     9100    10.00000
#> 83     9200    10.00000
#> 84     9200    10.00000
#> 85     9300    10.00000
#> 86     9200    10.00000
#> 87     9000    10.00000
#> 88     9200    10.00000
#> 89     9300    10.00000
#> 90     9000    10.00000
#> 91     9400    10.00000
#> 92     9100    10.00000
#> 93     9500    10.00000
#> 94     9300    10.00000
#> 95     9000    10.00000
#> 96     9900    10.00000
#> 97     9400    10.00000
#> 98     9300    10.00000
#> 99     9800    10.00000
#> 100    9500    10.00000
#> 101   71300    31.62278
#> 102   28600    31.62278
#> 103   25200    31.62278
#> 104   24800    31.62278
#> 105   27500    31.62278
#> 106   24700    31.62278
#> 107   23500    31.62278
#> 108   26800    31.62278
#> 109   25200    31.62278
#> 110   22400    31.62278
#> 111   22400    31.62278
#> 112   22200    31.62278
#> 113   21900    31.62278
#> 114   23500    31.62278
#> 115   22500    31.62278
#> 116   22200    31.62278
#> 117   21500    31.62278
#> 118   21600    31.62278
#> 119   21500    31.62278
#> 120   22400    31.62278
#> 121   22200    31.62278
#> 122   21500    31.62278
#> 123   23000    31.62278
#> 124   21600    31.62278
#> 125   21700    31.62278
#> 126   22100    31.62278
#> 127   22200    31.62278
#> 128   22400    31.62278
#> 129   21300    31.62278
#> 130   22200    31.62278
#> 131   21800    31.62278
#> 132   21500    31.62278
#> 133   21900    31.62278
#> 134   22400    31.62278
#> 135   21500    31.62278
#> 136   22000    31.62278
#> 137   21700    31.62278
#> 138   23000    31.62278
#> 139   21500    31.62278
#> 140   21700    31.62278
#> 141   21900    31.62278
#> 142   21900    31.62278
#> 143   22100    31.62278
#> 144   21800    31.62278
#> 145   22000    31.62278
#> 146   21800    31.62278
#> 147   21700    31.62278
#> 148   22400    31.62278
#> 149   27500    31.62278
#> 150   22500    31.62278
#> 151   22000    31.62278
#> 152   22200    31.62278
#> 153   21600    31.62278
#> 154   21300    31.62278
#> 155   21100    31.62278
#> 156   21900    31.62278
#> 157   21500    31.62278
#> 158   21800    31.62278
#> 159   21400    31.62278
#> 160   21500    31.62278
#> 161   22200    31.62278
#> 162   22000    31.62278
#> 163   21100    31.62278
#> 164   21200    31.62278
#> 165   21500    31.62278
#> 166   21300    31.62278
#> 167   21200    31.62278
#> 168   21400    31.62278
#> 169   21100    31.62278
#> 170   21200    31.62278
#> 171   21600    31.62278
#> 172   21200    31.62278
#> 173   22200    31.62278
#> 174   21500    31.62278
#> 175   21400    31.62278
#> 176   21500    31.62278
#> 177   21200    31.62278
#> 178   21500    31.62278
#> 179   21700    31.62278
#> 180   21900    31.62278
#> 181   21400    31.62278
#> 182   21800    31.62278
#> 183   21400    31.62278
#> 184   21100    31.62278
#> 185   22400    31.62278
#> 186   21700    31.62278
#> 187   21100    31.62278
#> 188   21300    31.62278
#> 189   21700    31.62278
#> 190   21500    31.62278
#> 191   21000    31.62278
#> 192   21800    31.62278
#> 193   21200    31.62278
#> 194   21500    31.62278
#> 195   25200    31.62278
#> 196   22200    31.62278
#> 197   21500    31.62278
#> 198   21300    31.62278
#> 199   21400    31.62278
#> 200   23600    31.62278
#> 201   82100   100.00000
#> 202   49800   100.00000
#> 203   46300   100.00000
#> 204   46300   100.00000
#> 205   44900   100.00000
#> 206   43600   100.00000
#> 207   47000   100.00000
#> 208   44500   100.00000
#> 209   44400   100.00000
#> 210   44600   100.00000
#> 211   44900   100.00000
#> 212   45400   100.00000
#> 213   43800   100.00000
#> 214   51300   100.00000
#> 215   44000   100.00000
#> 216   43100   100.00000
#> 217   42800   100.00000
#> 218   43400   100.00000
#> 219   43200   100.00000
#> 220   43100   100.00000
#> 221   43800   100.00000
#> 222   43900   100.00000
#> 223   43900   100.00000
#> 224   43700   100.00000
#> 225   45800   100.00000
#> 226   47100   100.00000
#> 227   44000   100.00000
#> 228   43900   100.00000
#> 229   43300   100.00000
#> 230   43100   100.00000
#> 231   44000   100.00000
#> 232   43800   100.00000
#> 233   43000   100.00000
#> 234   44000   100.00000
#> 235   43300   100.00000
#> 236   50300   100.00000
#> 237   44700   100.00000
#> 238   45100   100.00000
#> 239   44300   100.00000
#> 240   44400   100.00000
#> 241   43800   100.00000
#> 242   44700   100.00000
#> 243   43000   100.00000
#> 244   49400   100.00000
#> 245   44900   100.00000
#> 246   43600   100.00000
#> 247   43400   100.00000
#> 248   43700   100.00000
#> 249   43500   100.00000
#> 250   43000   100.00000
#> 251   43500   100.00000
#> 252   44200   100.00000
#> 253   43200   100.00000
#> 254   44100   100.00000
#> 255   44200   100.00000
#> 256   44300   100.00000
#> 257   43900   100.00000
#> 258   50600   100.00000
#> 259   44700   100.00000
#> 260   45400   100.00000
#> 261   45100   100.00000
#> 262   44400   100.00000
#> 263   44900   100.00000
#> 264   42900   100.00000
#> 265   43200   100.00000
#> 266   42900   100.00000
#> 267   43600   100.00000
#> 268   43400   100.00000
#> 269   43200   100.00000
#> 270   42800   100.00000
#> 271   43300   100.00000
#> 272   43700   100.00000
#> 273   43500   100.00000
#> 274   43100   100.00000
#> 275   43000   100.00000
#> 276   43900   100.00000
#> 277   43400   100.00000
#> 278   43400   100.00000
#> 279   43400   100.00000
#> 280   43300   100.00000
#> 281   49700   100.00000
#> 282   45500   100.00000
#> 283   43000   100.00000
#> 284   43100   100.00000
#> 285   43700   100.00000
#> 286   43100   100.00000
#> 287   44000   100.00000
#> 288   43900   100.00000
#> 289   43200   100.00000
#> 290   42500   100.00000
#> 291   43900   100.00000
#> 292   43700   100.00000
#> 293   43400   100.00000
#> 294   44000   100.00000
#> 295   43400   100.00000
#> 296   43600   100.00000
#> 297   43600   100.00000
#> 298   44200   100.00000
#> 299   43300   100.00000
#> 300   43300   100.00000
#> 301   98800   316.22777
#> 302   69100   316.22777
#> 303   64800   316.22777
#> 304   63800   316.22777
#> 305   63600   316.22777
#> 306   63200   316.22777
#> 307   63700   316.22777
#> 308   70900   316.22777
#> 309   63900   316.22777
#> 310   64700   316.22777
#> 311   63900   316.22777
#> 312   63300   316.22777
#> 313   62200   316.22777
#> 314   62700   316.22777
#> 315   62900   316.22777
#> 316   62300   316.22777
#> 317   63000   316.22777
#> 318   63400   316.22777
#> 319   63200   316.22777
#> 320   63600   316.22777
#> 321   63500   316.22777
#> 322   64000   316.22777
#> 323   69400   316.22777
#> 324   64300   316.22777
#> 325   63300   316.22777
#> 326   64300   316.22777
#> 327   63600   316.22777
#> 328   63600   316.22777
#> 329   65300   316.22777
#> 330   62200   316.22777
#> 331   62500   316.22777
#> 332   63800   316.22777
#> 333   62700   316.22777
#> 334   63700   316.22777
#> 335   63100   316.22777
#> 336   63200   316.22777
#> 337   63500   316.22777
#> 338   64300   316.22777
#> 339   69900   316.22777
#> 340   64100   316.22777
#> 341   64100   316.22777
#> 342   63300   316.22777
#> 343   63500   316.22777
#> 344   64400   316.22777
#> 345   63700   316.22777
#> 346   63200   316.22777
#> 347   63200   316.22777
#> 348   62600   316.22777
#> 349   63700   316.22777
#> 350   63200   316.22777
#> 351   64200   316.22777
#> 352   63600   316.22777
#> 353   63200   316.22777
#> 354   65500   316.22777
#> 355   67100   316.22777
#> 356   63400   316.22777
#> 357   62100   316.22777
#> 358   63200   316.22777
#> 359   62200   316.22777
#> 360   63100   316.22777
#> 361   62800   316.22777
#> 362   62600   316.22777
#> 363   62400   316.22777
#> 364   63400   316.22777
#> 365   63100   316.22777
#> 366   62500   316.22777
#> 367   62500   316.22777
#> 368   63100   316.22777
#> 369   62500   316.22777
#> 370   68800   316.22777
#> 371   64200   316.22777
#> 372   62500   316.22777
#> 373   62600   316.22777
#> 374   63700   316.22777
#> 375   63500   316.22777
#> 376   62600   316.22777
#> 377   63100   316.22777
#> 378   63500   316.22777
#> 379   63000   316.22777
#> 380   62400   316.22777
#> 381   63100   316.22777
#> 382   61900   316.22777
#> 383   62300   316.22777
#> 384   62200   316.22777
#> 385   62600   316.22777
#> 386   69900   316.22777
#> 387   63300   316.22777
#> 388   62900   316.22777
#> 389   62300   316.22777
#> 390   64200   316.22777
#> 391   63700   316.22777
#> 392   62600   316.22777
#> 393   63400   316.22777
#> 394   62700   316.22777
#> 395   63100   316.22777
#> 396   63500   316.22777
#> 397   63100   316.22777
#> 398   63100   316.22777
#> 399   62300   316.22777
#> 400   62700   316.22777
#> 401  274700  1000.00000
#> 402  232500  1000.00000
#> 403  228300  1000.00000
#> 404  230300  1000.00000
#> 405  226700  1000.00000
#> 406  235200  1000.00000
#> 407  228000  1000.00000
#> 408  233300  1000.00000
#> 409  229000  1000.00000
#> 410  235000  1000.00000
#> 411  228500  1000.00000
#> 412  229300  1000.00000
#> 413  234700  1000.00000
#> 414  229600  1000.00000
#> 415  227900  1000.00000
#> 416  229800  1000.00000
#> 417  231000  1000.00000
#> 418  232400  1000.00000
#> 419  229200  1000.00000
#> 420  228600  1000.00000
#> 421  233900  1000.00000
#> 422  228800  1000.00000
#> 423  231200  1000.00000
#> 424  226600  1000.00000
#> 425  234000  1000.00000
#> 426  229400  1000.00000
#> 427  227900  1000.00000
#> 428  230100  1000.00000
#> 429  228400  1000.00000
#> 430  268900  1000.00000
#> 431  240700  1000.00000
#> 432  227600  1000.00000
#> 433  155700  1000.00000
#> 434  161400  1000.00000
#> 435  158800  1000.00000
#> 436  138400  1000.00000
#> 437  136400  1000.00000
#> 438  134900  1000.00000
#> 439  135300  1000.00000
#> 440  134600  1000.00000
#> 441  133800  1000.00000
#> 442  192800  1000.00000
#> 443  217400  1000.00000
#> 444  204900  1000.00000
#> 445  205100  1000.00000
#> 446  203800  1000.00000
#> 447  222100  1000.00000
#> 448  205500  1000.00000
#> 449  205300  1000.00000
#> 450  203600  1000.00000
#> 451  229700  1000.00000
#> 452  227500  1000.00000
#> 453  209600  1000.00000
#> 454  204900  1000.00000
#> 455  204200  1000.00000
#> 456  258300  1000.00000
#> 457  298200  1000.00000
#> 458  296000  1000.00000
#> 459  292300  1000.00000
#> 460  229700  1000.00000
#> 461  228400  1000.00000
#> 462  227500  1000.00000
#> 463  229700  1000.00000
#> 464  233400  1000.00000
#> 465  228100  1000.00000
#> 466  227700  1000.00000
#> 467  227800  1000.00000
#> 468  288300  1000.00000
#> 469  302000  1000.00000
#> 470  295100  1000.00000
#> 471  299200  1000.00000
#> 472  230200  1000.00000
#> 473  228300  1000.00000
#> 474  229000  1000.00000
#> 475  232200  1000.00000
#> 476  229200  1000.00000
#> 477  226300  1000.00000
#> 478  227400  1000.00000
#> 479  227200  1000.00000
#> 480  239500  1000.00000
#> 481  232600  1000.00000
#> 482  230700  1000.00000
#> 483  232700  1000.00000
#> 484  235600  1000.00000
#> 485  229700  1000.00000
#> 486  234200  1000.00000
#> 487  229500  1000.00000
#> 488  230700  1000.00000
#> 489  236100  1000.00000
#> 490  228000  1000.00000
#> 491  229000  1000.00000
#> 492  229300  1000.00000
#> 493  232500  1000.00000
#> 494  228200  1000.00000
#> 495  228000  1000.00000
#> 496  228000  1000.00000
#> 497  232300  1000.00000
#> 498  229600  1000.00000
#> 499  228500  1000.00000
#> 500  229300  1000.00000
#> 501  676500  3162.27766
#> 502  418900  3162.27766
#> 503  414700  3162.27766
#> 504  663400  3162.27766
#> 505  640700  3162.27766
#> 506  634200  3162.27766
#> 507  647400  3162.27766
#> 508  647200  3162.27766
#> 509  640600  3162.27766
#> 510  644300  3162.27766
#> 511  639900  3162.27766
#> 512  639400  3162.27766
#> 513  642400  3162.27766
#> 514  654000  3162.27766
#> 515  653300  3162.27766
#> 516  649000  3162.27766
#> 517  643500  3162.27766
#> 518  647200  3162.27766
#> 519  653400  3162.27766
#> 520  643600  3162.27766
#> 521  640500  3162.27766
#> 522  641500  3162.27766
#> 523  639600  3162.27766
#> 524  657400  3162.27766
#> 525  644300  3162.27766
#> 526  636100  3162.27766
#> 527  664600  3162.27766
#> 528  647100  3162.27766
#> 529  630600  3162.27766
#> 530  656200  3162.27766
#> 531  692300  3162.27766
#> 532  927300  3162.27766
#> 533  917500  3162.27766
#> 534  701200  3162.27766
#> 535  701900  3162.27766
#> 536  707800  3162.27766
#> 537  707900  3162.27766
#> 538  702200  3162.27766
#> 539  700200  3162.27766
#> 540  700300  3162.27766
#> 541  712300  3162.27766
#> 542  936700  3162.27766
#> 543  824600  3162.27766
#> 544  799300  3162.27766
#> 545  563800  3162.27766
#> 546  434000  3162.27766
#> 547  418000  3162.27766
#> 548  446200  3162.27766
#> 549  632000  3162.27766
#> 550  512000  3162.27766
#> 551  415100  3162.27766
#> 552  463800  3162.27766
#> 553  628200  3162.27766
#> 554  502200  3162.27766
#> 555  421000  3162.27766
#> 556  460500  3162.27766
#> 557  667200  3162.27766
#> 558  440000  3162.27766
#> 559  412400  3162.27766
#> 560  418200  3162.27766
#> 561  417000  3162.27766
#> 562  413900  3162.27766
#> 563  418800  3162.27766
#> 564  410900  3162.27766
#> 565  420300  3162.27766
#> 566  416900  3162.27766
#> 567  420400  3162.27766
#> 568  416100  3162.27766
#> 569  414300  3162.27766
#> 570  421400  3162.27766
#> 571  414300  3162.27766
#> 572  417000  3162.27766
#> 573  413700  3162.27766
#> 574  418100  3162.27766
#> 575  418800  3162.27766
#> 576  414500  3162.27766
#> 577  421400  3162.27766
#> 578  416500  3162.27766
#> 579  421100  3162.27766
#> 580  420500  3162.27766
#> 581  416300  3162.27766
#> 582  791100  3162.27766
#> 583  777200  3162.27766
#> 584  703700  3162.27766
#> 585  698900  3162.27766
#> 586  703600  3162.27766
#> 587  722400  3162.27766
#> 588  701000  3162.27766
#> 589  700700  3162.27766
#> 590  702400  3162.27766
#> 591  703100  3162.27766
#> 592  699800  3162.27766
#> 593  700700  3162.27766
#> 594  706200  3162.27766
#> 595  704500  3162.27766
#> 596  705300  3162.27766
#> 597  702400  3162.27766
#> 598  703400  3162.27766
#> 599  701100  3162.27766
#> 600  701400  3162.27766
#> 601 1453600 10000.00000
#> 602 1340600 10000.00000
#> 603 1379500 10000.00000
#> 604 1328200 10000.00000
#> 605 1312000 10000.00000
#> 606 1305000 10000.00000
#> 607 1378400 10000.00000
#> 608 1302900 10000.00000
#> 609 1305200 10000.00000
#> 610 1306600 10000.00000
#> 611 1304800 10000.00000
#> 612 1304700 10000.00000
#> 613 1341300 10000.00000
#> 614 1312000 10000.00000
#> 615 1380600 10000.00000
#> 616 1300200 10000.00000
#> 617 1304300 10000.00000
#> 618 2232800 10000.00000
#> 619 2208700 10000.00000
#> 620 2201400 10000.00000
#> 621 2195400 10000.00000
#> 622 2190400 10000.00000
#> 623 2195100 10000.00000
#> 624 2203700 10000.00000
#> 625 2184500 10000.00000
#> 626 2195900 10000.00000
#> 627 2193900 10000.00000
#> 628 2196000 10000.00000
#> 629 2201800 10000.00000
#> 630 2189700 10000.00000
#> 631 2200300 10000.00000
#> 632 2195800 10000.00000
#> 633 2204400 10000.00000
#> 634 2206900 10000.00000
#> 635 2214100 10000.00000
#> 636 2198900 10000.00000
#> 637 2190100 10000.00000
#> 638 2197400 10000.00000
#> 639 2198600 10000.00000
#> 640 2209800 10000.00000
#> 641 2205300 10000.00000
#> 642 2213100 10000.00000
#> 643 2217300 10000.00000
#> 644 2197000 10000.00000
#> 645 2190700 10000.00000
#> 646 2201800 10000.00000
#> 647 2193900 10000.00000
#> 648 2194400 10000.00000
#> 649 2191900 10000.00000
#> 650 2194700 10000.00000
#> 651 2201300 10000.00000
#> 652 2199600 10000.00000
#> 653 2208400 10000.00000
#> 654 2200700 10000.00000
#> 655 2198300 10000.00000
#> 656 2192400 10000.00000
#> 657 2190600 10000.00000
#> 658 2232700 10000.00000
#> 659 2212900 10000.00000
#> 660 2215900 10000.00000
#> 661 2195200 10000.00000
#> 662 2194800 10000.00000
#> 663 2202100 10000.00000
#> 664 2210700 10000.00000
#> 665 2205200 10000.00000
#> 666 2203600 10000.00000
#> 667 2222900 10000.00000
#> 668 2177400 10000.00000
#> 669 2223200 10000.00000
#> 670 2514900 10000.00000
#> 671 2255400 10000.00000
#> 672 2224300 10000.00000
#> 673 2255600 10000.00000
#> 674 2270800 10000.00000
#> 675 1530500 10000.00000
#> 676 1301600 10000.00000
#> 677 1298900 10000.00000
#> 678 1298700 10000.00000
#> 679 1673500 10000.00000
#> 680 2441400 10000.00000
#> 681 2200600 10000.00000
#> 682 2192200 10000.00000
#> 683 1834100 10000.00000
#> 684 1885600 10000.00000
#> 685 2522900 10000.00000
#> 686 2358300 10000.00000
#> 687 2348200 10000.00000
#> 688 2348800 10000.00000
#> 689 2353000 10000.00000
#> 690 2099800 10000.00000
#> 691 1997500 10000.00000
#> 692 2414000 10000.00000
#> 693 2337600 10000.00000
#> 694 2335000 10000.00000
#> 695 2328100 10000.00000
#> 696 2336900 10000.00000
#> 697 2352500 10000.00000
#> 698 2343100 10000.00000
#> 699 2324100 10000.00000
#> 700 2385100 10000.00000