Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the quick sort algorithm (with sampling performed)
# against a set of increasing input data sizes:
input.sizes = 10^seq(1, 3, by = 0.5)
asymptoticTimings(sort(sample(1:100, data.sizes, replace = TRUE), method = "quick"), input.sizes)

#>     Timings Data sizes
#> 1    180601   10.00000
#> 2     92002   10.00000
#> 3     87601   10.00000
#> 4     80401   10.00000
#> 5     80201   10.00000
#> 6     90000   10.00000
#> 7     85301   10.00000
#> 8     77201   10.00000
#> 9     81100   10.00000
#> 10    81301   10.00000
#> 11    81001   10.00000
#> 12    74501   10.00000
#> 13    82302   10.00000
#> 14    72801   10.00000
#> 15    69500   10.00000
#> 16    85301   10.00000
#> 17    81201   10.00000
#> 18    77200   10.00000
#> 19    83202   10.00000
#> 20    75801   10.00000
#> 21    84801   10.00000
#> 22    74700   10.00000
#> 23    78801   10.00000
#> 24    77901   10.00000
#> 25    67401   10.00000
#> 26    80301   10.00000
#> 27    85901   10.00000
#> 28    75600   10.00000
#> 29    73102   10.00000
#> 30    77301   10.00000
#> 31    84602   10.00000
#> 32    72300   10.00000
#> 33    67101   10.00000
#> 34    70100   10.00000
#> 35    75901   10.00000
#> 36    66301   10.00000
#> 37    82801   10.00000
#> 38    82800   10.00000
#> 39    75301   10.00000
#> 40    77601   10.00000
#> 41    76402   10.00000
#> 42    84901   10.00000
#> 43    75701   10.00000
#> 44    77701   10.00000
#> 45    80301   10.00000
#> 46    69700   10.00000
#> 47    76301   10.00000
#> 48    83901   10.00000
#> 49    77301   10.00000
#> 50    72101   10.00000
#> 51    76501   10.00000
#> 52    75301   10.00000
#> 53    82101   10.00000
#> 54    74601   10.00000
#> 55    77301   10.00000
#> 56    76101   10.00000
#> 57    71601   10.00000
#> 58    81501   10.00000
#> 59    87100   10.00000
#> 60    77402   10.00000
#> 61    88701   10.00000
#> 62    78400   10.00000
#> 63    78201   10.00000
#> 64    80801   10.00000
#> 65    73501   10.00000
#> 66    77402   10.00000
#> 67    74700   10.00000
#> 68    62401   10.00000
#> 69    70201   10.00000
#> 70    87500   10.00000
#> 71    94002   10.00000
#> 72    74001   10.00000
#> 73    81300   10.00000
#> 74    72802   10.00000
#> 75    85701   10.00000
#> 76    74802   10.00000
#> 77    76801   10.00000
#> 78    76902   10.00000
#> 79    63901   10.00000
#> 80    69202   10.00000
#> 81    85601   10.00000
#> 82    88602   10.00000
#> 83    73501   10.00000
#> 84    77101   10.00000
#> 85    76900   10.00000
#> 86    82401   10.00000
#> 87    76601   10.00000
#> 88    73101   10.00000
#> 89    79701   10.00000
#> 90    69601   10.00000
#> 91    63501   10.00000
#> 92    83101   10.00000
#> 93    94700   10.00000
#> 94    74802   10.00000
#> 95    81201   10.00000
#> 96    77801   10.00000
#> 97    79301   10.00000
#> 98    81801   10.00000
#> 99    75302   10.00000
#> 100   78701   10.00000
#> 101  175201   31.62278
#> 102   95001   31.62278
#> 103   82800   31.62278
#> 104   91201   31.62278
#> 105   79501   31.62278
#> 106   84100   31.62278
#> 107   81401   31.62278
#> 108   65401   31.62278
#> 109   73201   31.62278
#> 110   88301   31.62278
#> 111   80101   31.62278
#> 112   76001   31.62278
#> 113   82701   31.62278
#> 114   77100   31.62278
#> 115   83401   31.62278
#> 116   74501   31.62278
#> 117   77802   31.62278
#> 118   75001   31.62278
#> 119   71000   31.62278
#> 120   83101   31.62278
#> 121   86000   31.62278
#> 122   74901   31.62278
#> 123   78101   31.62278
#> 124   78301   31.62278
#> 125   82601   31.62278
#> 126   76301   31.62278
#> 127   75601   31.62278
#> 128   80001   31.62278
#> 129   68601   31.62278
#> 130   79201   31.62278
#> 131   85801   31.62278
#> 132   78801   31.62278
#> 133   77701   31.62278
#> 134   79701   31.62278
#> 135   82402   31.62278
#> 136   76702   31.62278
#> 137   74101   31.62278
#> 138   80900   31.62278
#> 139   72001   31.62278
#> 140   77200   31.62278
#> 141   85102   31.62278
#> 142   77701   31.62278
#> 143   74202   31.62278
#> 144   80501   31.62278
#> 145   81201   31.62278
#> 146   80701   31.62278
#> 147   73700   31.62278
#> 148   77701   31.62278
#> 149   74301   31.62278
#> 150   67401   31.62278
#> 151   82801   31.62278
#> 152   80901   31.62278
#> 153   74401   31.62278
#> 154   77601   31.62278
#> 155   76300   31.62278
#> 156   84301   31.62278
#> 157   78001   31.62278
#> 158   76701   31.62278
#> 159   79001   31.62278
#> 160   68101   31.62278
#> 161   78902   31.62278
#> 162   85901   31.62278
#> 163   74001   31.62278
#> 164   77201   31.62278
#> 165   77501   31.62278
#> 166   81700   31.62278
#> 167   76801   31.62278
#> 168   75601   31.62278
#> 169   80600   31.62278
#> 170   73201   31.62278
#> 171   79801   31.62278
#> 172   85600   31.62278
#> 173   75101   31.62278
#> 174   73901   31.62278
#> 175   77901   31.62278
#> 176   79302   31.62278
#> 177   78201   31.62278
#> 178   72501   31.62278
#> 179   79201   31.62278
#> 180   73001   31.62278
#> 181   65401   31.62278
#> 182   83802   31.62278
#> 183   85200   31.62278
#> 184   72202   31.62278
#> 185   72401   31.62278
#> 186   76901   31.62278
#> 187   81702   31.62278
#> 188   76101   31.62278
#> 189   74600   31.62278
#> 190   79800   31.62278
#> 191   66901   31.62278
#> 192   72401   31.62278
#> 193   84700   31.62278
#> 194   78801   31.62278
#> 195   73201   31.62278
#> 196   83001   31.62278
#> 197   75101   31.62278
#> 198   83802   31.62278
#> 199   74001   31.62278
#> 200   77401   31.62278
#> 201  167201  100.00000
#> 202  112900  100.00000
#> 203   86001  100.00000
#> 204   75401  100.00000
#> 205   75000  100.00000
#> 206   75202  100.00000
#> 207   85901  100.00000
#> 208   76702  100.00000
#> 209   80401  100.00000
#> 210   77801  100.00000
#> 211   64102  100.00000
#> 212   71801  100.00000
#> 213   93901  100.00000
#> 214   78001  100.00000
#> 215   72300  100.00000
#> 216   78401  100.00000
#> 217   75101  100.00000
#> 218   81302  100.00000
#> 219   72400  100.00000
#> 220   88400  100.00000
#> 221   72702  100.00000
#> 222   62801  100.00000
#> 223   74901  100.00000
#> 224   84501  100.00000
#> 225   70900  100.00000
#> 226   72801  100.00000
#> 227   74101  100.00000
#> 228   77101  100.00000
#> 229   72501  100.00000
#> 230   71200  100.00000
#> 231   75702  100.00000
#> 232   65201  100.00000
#> 233   67801  100.00000
#> 234   80900  100.00000
#> 235   75602  100.00000
#> 236   69301  100.00000
#> 237   76601  100.00000
#> 238   71601  100.00000
#> 239   77202  100.00000
#> 240   69501  100.00000
#> 241   81702  100.00000
#> 242  101701  100.00000
#> 243   50300  100.00000
#> 244   41801  100.00000
#> 245   38502  100.00000
#> 246   37001  100.00000
#> 247   36201  100.00000
#> 248   36301  100.00000
#> 249   41101  100.00000
#> 250   35700  100.00000
#> 251   36001  100.00000
#> 252   35201  100.00000
#> 253   47701  100.00000
#> 254   56301  100.00000
#> 255   43901  100.00000
#> 256   38601  100.00000
#> 257   42202  100.00000
#> 258   50701  100.00000
#> 259   57801  100.00000
#> 260   58801  100.00000
#> 261   43601  100.00000
#> 262   35602  100.00000
#> 263   35001  100.00000
#> 264   35200  100.00000
#> 265   34801  100.00000
#> 266   34901  100.00000
#> 267   35502  100.00000
#> 268   34801  100.00000
#> 269   35101  100.00000
#> 270   46801  100.00000
#> 271   60601  100.00000
#> 272   60702  100.00000
#> 273   39001  100.00000
#> 274   39301  100.00000
#> 275   49201  100.00000
#> 276   36500  100.00000
#> 277   34901  100.00000
#> 278   34601  100.00000
#> 279   34601  100.00000
#> 280   34401  100.00000
#> 281   34601  100.00000
#> 282   35401  100.00000
#> 283   34901  100.00000
#> 284   35901  100.00000
#> 285   34800  100.00000
#> 286   35600  100.00000
#> 287   53601  100.00000
#> 288   51101  100.00000
#> 289   41901  100.00000
#> 290   39101  100.00000
#> 291   48501  100.00000
#> 292   45202  100.00000
#> 293   38500  100.00000
#> 294   35801  100.00000
#> 295   35701  100.00000
#> 296   34601  100.00000
#> 297   46800  100.00000
#> 298   33500  100.00000
#> 299   34101  100.00000
#> 300   34801  100.00000
#> 301  198301  316.22777
#> 302   67801  316.22777
#> 303   46500  316.22777
#> 304   41701  316.22777
#> 305   36101  316.22777
#> 306   35101  316.22777
#> 307   33500  316.22777
#> 308   33301  316.22777
#> 309   33701  316.22777
#> 310   32201  316.22777
#> 311   32500  316.22777
#> 312   33100  316.22777
#> 313   32800  316.22777
#> 314   33200  316.22777
#> 315   32501  316.22777
#> 316   32500  316.22777
#> 317   32501  316.22777
#> 318   32501  316.22777
#> 319   32901  316.22777
#> 320   32800  316.22777
#> 321   40300  316.22777
#> 322   33901  316.22777
#> 323   33001  316.22777
#> 324   33002  316.22777
#> 325   32401  316.22777
#> 326   32601  316.22777
#> 327   33000  316.22777
#> 328   33001  316.22777
#> 329   32601  316.22777
#> 330   32901  316.22777
#> 331   32301  316.22777
#> 332   32901  316.22777
#> 333   32701  316.22777
#> 334   33601  316.22777
#> 335   32600  316.22777
#> 336   32101  316.22777
#> 337   32301  316.22777
#> 338   33500  316.22777
#> 339   32600  316.22777
#> 340   32601  316.22777
#> 341   32901  316.22777
#> 342   32801  316.22777
#> 343   32401  316.22777
#> 344   33401  316.22777
#> 345   32901  316.22777
#> 346   32501  316.22777
#> 347   32502  316.22777
#> 348   33701  316.22777
#> 349   32601  316.22777
#> 350   33302  316.22777
#> 351   55301  316.22777
#> 352   41401  316.22777
#> 353   34801  316.22777
#> 354   33901  316.22777
#> 355   32801  316.22777
#> 356   33001  316.22777
#> 357   32901  316.22777
#> 358   33801  316.22777
#> 359   32401  316.22777
#> 360   33202  316.22777
#> 361   32701  316.22777
#> 362   33401  316.22777
#> 363   32500  316.22777
#> 364   33301  316.22777
#> 365   32601  316.22777
#> 366   32301  316.22777
#> 367   32601  316.22777
#> 368   32501  316.22777
#> 369   33001  316.22777
#> 370   33300  316.22777
#> 371   33301  316.22777
#> 372   32701  316.22777
#> 373   33301  316.22777
#> 374   32702  316.22777
#> 375   32401  316.22777
#> 376   32801  316.22777
#> 377   32701  316.22777
#> 378   32500  316.22777
#> 379   32501  316.22777
#> 380   39801  316.22777
#> 381   33401  316.22777
#> 382   32901  316.22777
#> 383   32601  316.22777
#> 384   33101  316.22777
#> 385   33101  316.22777
#> 386   32601  316.22777
#> 387   32100  316.22777
#> 388   32600  316.22777
#> 389   34301  316.22777
#> 390   32401  316.22777
#> 391   33001  316.22777
#> 392   33002  316.22777
#> 393   32802  316.22777
#> 394   33101  316.22777
#> 395   33001  316.22777
#> 396   32302  316.22777
#> 397   33001  316.22777
#> 398   32202  316.22777
#> 399   32800  316.22777
#> 400   32501  316.22777
#> 401  130601 1000.00000
#> 402   54401 1000.00000
#> 403   40400 1000.00000
#> 404   36401 1000.00000
#> 405   33901 1000.00000
#> 406   32401 1000.00000
#> 407   32101 1000.00000
#> 408   31502 1000.00000
#> 409   31701 1000.00000
#> 410   31301 1000.00000
#> 411   31202 1000.00000
#> 412   31801 1000.00000
#> 413   31401 1000.00000
#> 414   31601 1000.00000
#> 415   31802 1000.00000
#> 416   31302 1000.00000
#> 417   30901 1000.00000
#> 418   31601 1000.00000
#> 419   31501 1000.00000
#> 420   31901 1000.00000
#> 421   31802 1000.00000
#> 422   31701 1000.00000
#> 423   31701 1000.00000
#> 424   31802 1000.00000
#> 425   30900 1000.00000
#> 426   31300 1000.00000
#> 427   31400 1000.00000
#> 428   32901 1000.00000
#> 429   31202 1000.00000
#> 430   31201 1000.00000
#> 431   30901 1000.00000
#> 432   31202 1000.00000
#> 433   31301 1000.00000
#> 434   31901 1000.00000
#> 435   31501 1000.00000
#> 436   32101 1000.00000
#> 437   31701 1000.00000
#> 438   31901 1000.00000
#> 439   31701 1000.00000
#> 440   31600 1000.00000
#> 441   32501 1000.00000
#> 442   32302 1000.00000
#> 443   31901 1000.00000
#> 444   31700 1000.00000
#> 445   31401 1000.00000
#> 446   32401 1000.00000
#> 447   31800 1000.00000
#> 448   32200 1000.00000
#> 449   31801 1000.00000
#> 450   31401 1000.00000
#> 451   31401 1000.00000
#> 452   31102 1000.00000
#> 453   31301 1000.00000
#> 454   31402 1000.00000
#> 455   31300 1000.00000
#> 456   32501 1000.00000
#> 457   31401 1000.00000
#> 458   32300 1000.00000
#> 459   31601 1000.00000
#> 460   32002 1000.00000
#> 461   32301 1000.00000
#> 462   32100 1000.00000
#> 463   48701 1000.00000
#> 464   33801 1000.00000
#> 465   31601 1000.00000
#> 466   31201 1000.00000
#> 467   31801 1000.00000
#> 468   31102 1000.00000
#> 469   47901 1000.00000
#> 470   41301 1000.00000
#> 471   33701 1000.00000
#> 472   32000 1000.00000
#> 473   31501 1000.00000
#> 474   31701 1000.00000
#> 475   32000 1000.00000
#> 476   31700 1000.00000
#> 477   33201 1000.00000
#> 478   31101 1000.00000
#> 479   31602 1000.00000
#> 480   31700 1000.00000
#> 481   31401 1000.00000
#> 482   31501 1000.00000
#> 483   30802 1000.00000
#> 484   32801 1000.00000
#> 485   32800 1000.00000
#> 486   32901 1000.00000
#> 487   31501 1000.00000
#> 488   32001 1000.00000
#> 489   33501 1000.00000
#> 490   31401 1000.00000
#> 491   31202 1000.00000
#> 492   31901 1000.00000
#> 493   31901 1000.00000
#> 494   31501 1000.00000
#> 495   32102 1000.00000
#> 496   31401 1000.00000
#> 497   31601 1000.00000
#> 498   31301 1000.00000
#> 499   33001 1000.00000
#> 500   32100 1000.00000