Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the quick sort algorithm (with sampling performed)
# against a set of increasing input data sizes:
input.sizes = 10^seq(1, 3, by = 0.5)
asymptoticTimings(sort(sample(1:100, data.sizes, replace = TRUE), method = "quick"), input.sizes)

#>     Timings Data sizes
#> 1    217601   10.00000
#> 2    128701   10.00000
#> 3     97501   10.00000
#> 4     88201   10.00000
#> 5     80701   10.00000
#> 6     93501   10.00000
#> 7     95301   10.00000
#> 8     58202   10.00000
#> 9     41801   10.00000
#> 10    37901   10.00000
#> 11    48700   10.00000
#> 12    38601   10.00000
#> 13    35502   10.00000
#> 14    35301   10.00000
#> 15    34601   10.00000
#> 16    34200   10.00000
#> 17    34401   10.00000
#> 18    34900   10.00000
#> 19    34401   10.00000
#> 20    34401   10.00000
#> 21    34301   10.00000
#> 22    34700   10.00000
#> 23    34600   10.00000
#> 24    34501   10.00000
#> 25    34200   10.00000
#> 26    36501   10.00000
#> 27    42101   10.00000
#> 28    34901   10.00000
#> 29    34502   10.00000
#> 30    36201   10.00000
#> 31    36301   10.00000
#> 32    34801   10.00000
#> 33    34902   10.00000
#> 34    38501   10.00000
#> 35    37802   10.00000
#> 36    35102   10.00000
#> 37    35601   10.00000
#> 38    41901   10.00000
#> 39    35201   10.00000
#> 40    34301   10.00000
#> 41    34401   10.00000
#> 42    35102   10.00000
#> 43    35001   10.00000
#> 44    34801   10.00000
#> 45    34301   10.00000
#> 46    35002   10.00000
#> 47    35001   10.00000
#> 48    34601   10.00000
#> 49    34301   10.00000
#> 50    34000   10.00000
#> 51    34701   10.00000
#> 52    35101   10.00000
#> 53    35300   10.00000
#> 54    35801   10.00000
#> 55    35501   10.00000
#> 56    35601   10.00000
#> 57    35601   10.00000
#> 58    40200   10.00000
#> 59    34401   10.00000
#> 60    34400   10.00000
#> 61    35301   10.00000
#> 62    35202   10.00000
#> 63    35100   10.00000
#> 64    34301   10.00000
#> 65    35201   10.00000
#> 66    42001   10.00000
#> 67    35500   10.00000
#> 68    34901   10.00000
#> 69    34902   10.00000
#> 70    34601   10.00000
#> 71    34901   10.00000
#> 72    34200   10.00000
#> 73    34301   10.00000
#> 74    34502   10.00000
#> 75    34701   10.00000
#> 76    34602   10.00000
#> 77    35901   10.00000
#> 78    36701   10.00000
#> 79    36201   10.00000
#> 80    34602   10.00000
#> 81    43401   10.00000
#> 82    34801   10.00000
#> 83    34801   10.00000
#> 84    34801   10.00000
#> 85    35801   10.00000
#> 86    34500   10.00000
#> 87    35100   10.00000
#> 88    34701   10.00000
#> 89    34301   10.00000
#> 90    35301   10.00000
#> 91    34202   10.00000
#> 92    35501   10.00000
#> 93    34301   10.00000
#> 94    60800   10.00000
#> 95    41901   10.00000
#> 96    37001   10.00000
#> 97    35502   10.00000
#> 98    36102   10.00000
#> 99    35401   10.00000
#> 100   35301   10.00000
#> 101  163901   31.62278
#> 102   79002   31.62278
#> 103   51301   31.62278
#> 104   45801   31.62278
#> 105   38901   31.62278
#> 106   38001   31.62278
#> 107   37500   31.62278
#> 108   37300   31.62278
#> 109   37000   31.62278
#> 110   36801   31.62278
#> 111   37400   31.62278
#> 112   37301   31.62278
#> 113   37401   31.62278
#> 114   37801   31.62278
#> 115   36701   31.62278
#> 116   37101   31.62278
#> 117   37401   31.62278
#> 118   36500   31.62278
#> 119   37901   31.62278
#> 120   36701   31.62278
#> 121   36801   31.62278
#> 122   36901   31.62278
#> 123   36901   31.62278
#> 124   49201   31.62278
#> 125   45401   31.62278
#> 126   39302   31.62278
#> 127   45700   31.62278
#> 128   37800   31.62278
#> 129   37701   31.62278
#> 130   38701   31.62278
#> 131   52001   31.62278
#> 132   42501   31.62278
#> 133  707801   31.62278
#> 134  656602   31.62278
#> 135  534501   31.62278
#> 136  575300   31.62278
#> 137  389002   31.62278
#> 138  204602   31.62278
#> 139  151401   31.62278
#> 140  181501   31.62278
#> 141   70400   31.62278
#> 142   58101   31.62278
#> 143   53201   31.62278
#> 144   80201   31.62278
#> 145   67701   31.62278
#> 146   77601   31.62278
#> 147   67801   31.62278
#> 148   69301   31.62278
#> 149   68401   31.62278
#> 150   68500   31.62278
#> 151   68501   31.62278
#> 152   69201   31.62278
#> 153   68002   31.62278
#> 154   68601   31.62278
#> 155   68501   31.62278
#> 156   63901   31.62278
#> 157   64501   31.62278
#> 158   71301   31.62278
#> 159   63001   31.62278
#> 160   62301   31.62278
#> 161   58900   31.62278
#> 162   57902   31.62278
#> 163   58801   31.62278
#> 164   58001   31.62278
#> 165   62001   31.62278
#> 166   59401   31.62278
#> 167   59801   31.62278
#> 168   61501   31.62278
#> 169   60101   31.62278
#> 170   61601   31.62278
#> 171   42601   31.62278
#> 172   43502   31.62278
#> 173   44900   31.62278
#> 174   38201   31.62278
#> 175   45801   31.62278
#> 176   71900   31.62278
#> 177   85602   31.62278
#> 178  712301   31.62278
#> 179  208401   31.62278
#> 180  104200   31.62278
#> 181   82101   31.62278
#> 182   98901   31.62278
#> 183   75602   31.62278
#> 184   71101   31.62278
#> 185   92101   31.62278
#> 186   69902   31.62278
#> 187   72801   31.62278
#> 188   77901   31.62278
#> 189   70301   31.62278
#> 190   42800   31.62278
#> 191   43001   31.62278
#> 192   68201   31.62278
#> 193   38801   31.62278
#> 194   34901   31.62278
#> 195   34301   31.62278
#> 196   53601   31.62278
#> 197   99801   31.62278
#> 198   86401   31.62278
#> 199   37601   31.62278
#> 200   55301   31.62278
#> 201  133701  100.00000
#> 202   57101  100.00000
#> 203   42702  100.00000
#> 204   40001  100.00000
#> 205   37301  100.00000
#> 206   42900  100.00000
#> 207   35801  100.00000
#> 208   35100  100.00000
#> 209   34701  100.00000
#> 210   34801  100.00000
#> 211   35000  100.00000
#> 212   34901  100.00000
#> 213   35201  100.00000
#> 214   35201  100.00000
#> 215   35602  100.00000
#> 216   35301  100.00000
#> 217   35100  100.00000
#> 218   35601  100.00000
#> 219   42701  100.00000
#> 220   35301  100.00000
#> 221   35101  100.00000
#> 222   34801  100.00000
#> 223   35001  100.00000
#> 224   35602  100.00000
#> 225   34801  100.00000
#> 226   37201  100.00000
#> 227   42201  100.00000
#> 228   35402  100.00000
#> 229   35000  100.00000
#> 230   36301  100.00000
#> 231   36001  100.00000
#> 232   34601  100.00000
#> 233   36400  100.00000
#> 234   46701  100.00000
#> 235   35101  100.00000
#> 236   35201  100.00000
#> 237   35102  100.00000
#> 238   35902  100.00000
#> 239   35401  100.00000
#> 240   35401  100.00000
#> 241   35801  100.00000
#> 242   39501  100.00000
#> 243   34802  100.00000
#> 244   34401  100.00000
#> 245   34801  100.00000
#> 246   35501  100.00000
#> 247   34901  100.00000
#> 248   34601  100.00000
#> 249   34801  100.00000
#> 250   34701  100.00000
#> 251   35301  100.00000
#> 252   36501  100.00000
#> 253   35802  100.00000
#> 254   35701  100.00000
#> 255   34402  100.00000
#> 256   36401  100.00000
#> 257   35001  100.00000
#> 258   35002  100.00000
#> 259   34601  100.00000
#> 260   34201  100.00000
#> 261   40902  100.00000
#> 262   35801  100.00000
#> 263   34701  100.00000
#> 264   34001  100.00000
#> 265   38201  100.00000
#> 266   36901  100.00000
#> 267   35002  100.00000
#> 268   34602  100.00000
#> 269   35001  100.00000
#> 270   34801  100.00000
#> 271   34701  100.00000
#> 272   34401  100.00000
#> 273   34301  100.00000
#> 274   34301  100.00000
#> 275   34400  100.00000
#> 276   34901  100.00000
#> 277   34501  100.00000
#> 278   37302  100.00000
#> 279   34901  100.00000
#> 280   34801  100.00000
#> 281   35101  100.00000
#> 282   34202  100.00000
#> 283   34401  100.00000
#> 284   35300  100.00000
#> 285   34700  100.00000
#> 286   35201  100.00000
#> 287   34801  100.00000
#> 288   37600  100.00000
#> 289   45901  100.00000
#> 290   35501  100.00000
#> 291   34501  100.00000
#> 292   36002  100.00000
#> 293   34801  100.00000
#> 294   35200  100.00000
#> 295   34901  100.00000
#> 296   34900  100.00000
#> 297   34601  100.00000
#> 298   35402  100.00000
#> 299   34901  100.00000
#> 300   35501  100.00000
#> 301  235200  316.22777
#> 302  118001  316.22777
#> 303  106501  316.22777
#> 304  112300  316.22777
#> 305  104501  316.22777
#> 306  123101  316.22777
#> 307  110202  316.22777
#> 308   88401  316.22777
#> 309   54600  316.22777
#> 310   44801  316.22777
#> 311   42501  316.22777
#> 312   41202  316.22777
#> 313   49201  316.22777
#> 314 1170801  316.22777
#> 315  189601  316.22777
#> 316  106901  316.22777
#> 317   55401  316.22777
#> 318   50000  316.22777
#> 319   49701  316.22777
#> 320   47001  316.22777
#> 321   45501  316.22777
#> 322   44701  316.22777
#> 323   43401  316.22777
#> 324   43101  316.22777
#> 325   42600  316.22777
#> 326   44001  316.22777
#> 327   48002  316.22777
#> 328   42101  316.22777
#> 329   41802  316.22777
#> 330   49101  316.22777
#> 331   41302  316.22777
#> 332   40702  316.22777
#> 333   85201  316.22777
#> 334   71301  316.22777
#> 335   77301  316.22777
#> 336   61101  316.22777
#> 337   61401  316.22777
#> 338   71700  316.22777
#> 339   68802  316.22777
#> 340   64401  316.22777
#> 341   61801  316.22777
#> 342   65500  316.22777
#> 343   69300  316.22777
#> 344   63301  316.22777
#> 345   61201  316.22777
#> 346   62801  316.22777
#> 347   78802  316.22777
#> 348   70902  316.22777
#> 349   94101  316.22777
#> 350   78201  316.22777
#> 351   69901  316.22777
#> 352   69601  316.22777
#> 353   73701  316.22777
#> 354   83401  316.22777
#> 355   74501  316.22777
#> 356   69801  316.22777
#> 357   73900  316.22777
#> 358   66501  316.22777
#> 359   66801  316.22777
#> 360   69002  316.22777
#> 361   70801  316.22777
#> 362   69800  316.22777
#> 363   81901  316.22777
#> 364   68301  316.22777
#> 365   55501  316.22777
#> 366   49400  316.22777
#> 367   41001  316.22777
#> 368   39801  316.22777
#> 369   39101  316.22777
#> 370   39101  316.22777
#> 371   38901  316.22777
#> 372   38500  316.22777
#> 373   38101  316.22777
#> 374   38601  316.22777
#> 375   38401  316.22777
#> 376   39001  316.22777
#> 377   39500  316.22777
#> 378   41501  316.22777
#> 379   38902  316.22777
#> 380   38302  316.22777
#> 381   38702  316.22777
#> 382   38200  316.22777
#> 383   38601  316.22777
#> 384   39000  316.22777
#> 385   38601  316.22777
#> 386   95901  316.22777
#> 387   72102  316.22777
#> 388   68501  316.22777
#> 389   70401  316.22777
#> 390   70901  316.22777
#> 391   70601  316.22777
#> 392   71001  316.22777
#> 393   70601  316.22777
#> 394   70101  316.22777
#> 395   68600  316.22777
#> 396   68300  316.22777
#> 397   70200  316.22777
#> 398   76201  316.22777
#> 399  111201  316.22777
#> 400   90201  316.22777
#> 401  172401 1000.00000
#> 402   87301 1000.00000
#> 403   66201 1000.00000
#> 404   63401 1000.00000
#> 405   64200 1000.00000
#> 406   62600 1000.00000
#> 407   71101 1000.00000
#> 408   66501 1000.00000
#> 409   64602 1000.00000
#> 410   57001 1000.00000
#> 411   62700 1000.00000
#> 412   53502 1000.00000
#> 413   58802 1000.00000
#> 414   58300 1000.00000
#> 415  105602 1000.00000
#> 416  107600 1000.00000
#> 417   83801 1000.00000
#> 418   69301 1000.00000
#> 419   67802 1000.00000
#> 420   66001 1000.00000
#> 421  139801 1000.00000
#> 422   75001 1000.00000
#> 423  108401 1000.00000
#> 424  115401 1000.00000
#> 425  134701 1000.00000
#> 426  110402 1000.00000
#> 427  172702 1000.00000
#> 428   83501 1000.00000
#> 429   69301 1000.00000
#> 430   71601 1000.00000
#> 431   69601 1000.00000
#> 432   64701 1000.00000
#> 433   67301 1000.00000
#> 434   63801 1000.00000
#> 435   56900 1000.00000
#> 436  189601 1000.00000
#> 437   77301 1000.00000
#> 438   77201 1000.00000
#> 439   69602 1000.00000
#> 440   64600 1000.00000
#> 441  131900 1000.00000
#> 442  164801 1000.00000
#> 443  219201 1000.00000
#> 444  171401 1000.00000
#> 445   86201 1000.00000
#> 446   72601 1000.00000
#> 447  108401 1000.00000
#> 448   91301 1000.00000
#> 449  124502 1000.00000
#> 450   79501 1000.00000
#> 451   87501 1000.00000
#> 452  117602 1000.00000
#> 453   76201 1000.00000
#> 454  112402 1000.00000
#> 455  104601 1000.00000
#> 456   75401 1000.00000
#> 457   80002 1000.00000
#> 458   78701 1000.00000
#> 459   75602 1000.00000
#> 460  126301 1000.00000
#> 461   57301 1000.00000
#> 462   46502 1000.00000
#> 463   41801 1000.00000
#> 464   56600 1000.00000
#> 465   77301 1000.00000
#> 466   71501 1000.00000
#> 467   72101 1000.00000
#> 468   69501 1000.00000
#> 469   66801 1000.00000
#> 470   68501 1000.00000
#> 471   71700 1000.00000
#> 472   68901 1000.00000
#> 473   68100 1000.00000
#> 474   67701 1000.00000
#> 475   66301 1000.00000
#> 476   64101 1000.00000
#> 477   64102 1000.00000
#> 478   69301 1000.00000
#> 479   63401 1000.00000
#> 480   61701 1000.00000
#> 481   63201 1000.00000
#> 482   65001 1000.00000
#> 483   65901 1000.00000
#> 484   70701 1000.00000
#> 485  217201 1000.00000
#> 486  148601 1000.00000
#> 487   71501 1000.00000
#> 488   60601 1000.00000
#> 489   63201 1000.00000
#> 490   87601 1000.00000
#> 491   72800 1000.00000
#> 492   90201 1000.00000
#> 493   92902 1000.00000
#> 494   64301 1000.00000
#> 495   61000 1000.00000
#> 496   60501 1000.00000
#> 497   62300 1000.00000
#> 498   60601 1000.00000
#> 499   61101 1000.00000
#> 500   81601 1000.00000