Asymptotic Timings Quantifying function

Function to compute benchmarked timings with different data sizes for an R expression

asymptoticTimings(e, data.sizes, max.seconds)

Arguments

e
An expression which is in the form of a function operating on 'N' (as the data size for the algorithm to be tested against for a run), which takes values from the used-supplied parameter data.sizes.

data.sizes
A vector/set of data sizes, which should preferably be a sequence in powers of ten, with mid-values included. Example: data.sizes = 10^seq(1, 4, by = 0.5)

max.seconds
The maximum number of seconds an iteration would be limited upto. (once the limit has been exceeded, further computations on incrementally larger dataset sizes won't be done) Optional, with the default value set to 1 second.

Value

A data frame comprising of the timings computed by microbenchmark and the corresponding dataset sizes.

Details

For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-function/

Examples

# Quantifying the runtimes for the quick sort algorithm (with sampling performed)
# against a set of increasing input data sizes:
input.sizes = 10^seq(1, 3, by = 0.5)
asymptoticTimings(sort(sample(1:100, data.sizes, replace = TRUE), method = "quick"), input.sizes)

#>     Timings Data sizes
#> 1    140900   10.00000
#> 2     57600   10.00000
#> 3     45000   10.00000
#> 4     41100   10.00000
#> 5     40700   10.00000
#> 6     39100   10.00000
#> 7     38900   10.00000
#> 8     39200   10.00000
#> 9     38500   10.00000
#> 10    38800   10.00000
#> 11    38600   10.00000
#> 12    38500   10.00000
#> 13    38700   10.00000
#> 14    38900   10.00000
#> 15    38500   10.00000
#> 16    39000   10.00000
#> 17    38200   10.00000
#> 18    38800   10.00000
#> 19    38500   10.00000
#> 20    38200   10.00000
#> 21    38100   10.00000
#> 22    38100   10.00000
#> 23    38600   10.00000
#> 24    38700   10.00000
#> 25    42400   10.00000
#> 26    44400   10.00000
#> 27    38500   10.00000
#> 28    38700   10.00000
#> 29    38000   10.00000
#> 30    38700   10.00000
#> 31    39500   10.00000
#> 32    38600   10.00000
#> 33    41600   10.00000
#> 34    56400   10.00000
#> 35    77600   10.00000
#> 36    70700   10.00000
#> 37    70200   10.00000
#> 38    69000   10.00000
#> 39    66800   10.00000
#> 40    77900   10.00000
#> 41    74100   10.00000
#> 42    81400   10.00000
#> 43    59400   10.00000
#> 44    43200   10.00000
#> 45    40200   10.00000
#> 46    39600   10.00000
#> 47    39000   10.00000
#> 48    39200   10.00000
#> 49    39900   10.00000
#> 50    44000   10.00000
#> 51    38500   10.00000
#> 52    39700   10.00000
#> 53    38900   10.00000
#> 54    62000   10.00000
#> 55    61700   10.00000
#> 56    65100   10.00000
#> 57    59400   10.00000
#> 58    69500   10.00000
#> 59    71200   10.00000
#> 60    67900   10.00000
#> 61    39700   10.00000
#> 62    41100   10.00000
#> 63    50400   10.00000
#> 64    40100   10.00000
#> 65    39600   10.00000
#> 66    38700   10.00000
#> 67    38500   10.00000
#> 68    38700   10.00000
#> 69    40300   10.00000
#> 70    35300   10.00000
#> 71    33600   10.00000
#> 72    33600   10.00000
#> 73    33100   10.00000
#> 74    32900   10.00000
#> 75    33600   10.00000
#> 76    33700   10.00000
#> 77    35700   10.00000
#> 78    32800   10.00000
#> 79    33100   10.00000
#> 80    33400   10.00000
#> 81    33700   10.00000
#> 82    33600   10.00000
#> 83    32700   10.00000
#> 84    33700   10.00000
#> 85    33200   10.00000
#> 86    32800   10.00000
#> 87    33100   10.00000
#> 88    32900   10.00000
#> 89    32800   10.00000
#> 90    33500   10.00000
#> 91    47900   10.00000
#> 92    40600   10.00000
#> 93    39000   10.00000
#> 94    38400   10.00000
#> 95    38700   10.00000
#> 96    38600   10.00000
#> 97    38400   10.00000
#> 98    38400   10.00000
#> 99    39100   10.00000
#> 100   38600   10.00000
#> 101  110000   31.62278
#> 102   50300   31.62278
#> 103   42400   31.62278
#> 104   39500   31.62278
#> 105   38300   31.62278
#> 106   38100   31.62278
#> 107   37400   31.62278
#> 108   37000   31.62278
#> 109   37500   31.62278
#> 110   36700   31.62278
#> 111   37600   31.62278
#> 112   37300   31.62278
#> 113   37500   31.62278
#> 114   36900   31.62278
#> 115   37400   31.62278
#> 116   37000   31.62278
#> 117   37300   31.62278
#> 118   37100   31.62278
#> 119   38400   31.62278
#> 120   42500   31.62278
#> 121   37000   31.62278
#> 122   36800   31.62278
#> 123   36900   31.62278
#> 124   37200   31.62278
#> 125   40800   31.62278
#> 126   37100   31.62278
#> 127   36900   31.62278
#> 128   37100   31.62278
#> 129   37200   31.62278
#> 130   37100   31.62278
#> 131   37500   31.62278
#> 132   37500   31.62278
#> 133   37800   31.62278
#> 134   37100   31.62278
#> 135   37200   31.62278
#> 136   37000   31.62278
#> 137   37200   31.62278
#> 138   36800   31.62278
#> 139   37500   31.62278
#> 140   37200   31.62278
#> 141   37200   31.62278
#> 142   37500   31.62278
#> 143   37000   31.62278
#> 144   37200   31.62278
#> 145   36500   31.62278
#> 146   57700   31.62278
#> 147   65900   31.62278
#> 148   40700   31.62278
#> 149   37700   31.62278
#> 150   35300   31.62278
#> 151   32300   31.62278
#> 152   31500   31.62278
#> 153   32100   31.62278
#> 154   32200   31.62278
#> 155   31900   31.62278
#> 156   32600   31.62278
#> 157   31700   31.62278
#> 158   31600   31.62278
#> 159   32100   31.62278
#> 160   31400   31.62278
#> 161   31000   31.62278
#> 162   31200   31.62278
#> 163   31000   31.62278
#> 164   31100   31.62278
#> 165   30900   31.62278
#> 166   32300   31.62278
#> 167   31500   31.62278
#> 168   32000   31.62278
#> 169   35500   31.62278
#> 170   32300   31.62278
#> 171   31500   31.62278
#> 172   31200   31.62278
#> 173   31300   31.62278
#> 174   30900   31.62278
#> 175   41700   31.62278
#> 176   38900   31.62278
#> 177   37900   31.62278
#> 178   38200   31.62278
#> 179   37300   31.62278
#> 180   37500   31.62278
#> 181   37600   31.62278
#> 182   36900   31.62278
#> 183   37300   31.62278
#> 184   37100   31.62278
#> 185   37500   31.62278
#> 186   37100   31.62278
#> 187   37600   31.62278
#> 188   37300   31.62278
#> 189   37400   31.62278
#> 190   37100   31.62278
#> 191   38100   31.62278
#> 192   37600   31.62278
#> 193   38100   31.62278
#> 194   37300   31.62278
#> 195   37000   31.62278
#> 196   37000   31.62278
#> 197   37100   31.62278
#> 198   37100   31.62278
#> 199   37000   31.62278
#> 200   37800   31.62278
#> 201  125200  100.00000
#> 202   46400  100.00000
#> 203   38600  100.00000
#> 204   35600  100.00000
#> 205   33400  100.00000
#> 206   32700  100.00000
#> 207   32200  100.00000
#> 208   31500  100.00000
#> 209   31700  100.00000
#> 210   32300  100.00000
#> 211   31900  100.00000
#> 212   31700  100.00000
#> 213   32000  100.00000
#> 214   31600  100.00000
#> 215   31400  100.00000
#> 216   45800  100.00000
#> 217   56200  100.00000
#> 218   35100  100.00000
#> 219   31900  100.00000
#> 220   32100  100.00000
#> 221   31700  100.00000
#> 222   31500  100.00000
#> 223   32200  100.00000
#> 224   32000  100.00000
#> 225   35700  100.00000
#> 226   31300  100.00000
#> 227   31600  100.00000
#> 228   31700  100.00000
#> 229   31100  100.00000
#> 230   31600  100.00000
#> 231   32000  100.00000
#> 232   31900  100.00000
#> 233   31500  100.00000
#> 234   31400  100.00000
#> 235   31400  100.00000
#> 236   31400  100.00000
#> 237   32200  100.00000
#> 238   31200  100.00000
#> 239   31400  100.00000
#> 240   31600  100.00000
#> 241   32200  100.00000
#> 242   31900  100.00000
#> 243   31400  100.00000
#> 244   31400  100.00000
#> 245   31100  100.00000
#> 246   38700  100.00000
#> 247   32500  100.00000
#> 248   31300  100.00000
#> 249   31500  100.00000
#> 250   32600  100.00000
#> 251   32500  100.00000
#> 252   31600  100.00000
#> 253   31700  100.00000
#> 254   31500  100.00000
#> 255   31600  100.00000
#> 256   31600  100.00000
#> 257   31400  100.00000
#> 258   31200  100.00000
#> 259   31600  100.00000
#> 260   31500  100.00000
#> 261   31100  100.00000
#> 262   31200  100.00000
#> 263   30600  100.00000
#> 264   31000  100.00000
#> 265   30800  100.00000
#> 266   31900  100.00000
#> 267   31200  100.00000
#> 268   31200  100.00000
#> 269   31200  100.00000
#> 270   32100  100.00000
#> 271   31100  100.00000
#> 272   30800  100.00000
#> 273   31300  100.00000
#> 274   31300  100.00000
#> 275   31400  100.00000
#> 276   32800  100.00000
#> 277   39200  100.00000
#> 278   33100  100.00000
#> 279   31500  100.00000
#> 280   31300  100.00000
#> 281   32000  100.00000
#> 282   31100  100.00000
#> 283   31400  100.00000
#> 284   31200  100.00000
#> 285   31600  100.00000
#> 286   31700  100.00000
#> 287   31000  100.00000
#> 288   31500  100.00000
#> 289   31400  100.00000
#> 290   30800  100.00000
#> 291   32500  100.00000
#> 292   31400  100.00000
#> 293   32000  100.00000
#> 294   30900  100.00000
#> 295   31500  100.00000
#> 296   31500  100.00000
#> 297   31300  100.00000
#> 298   31500  100.00000
#> 299   31200  100.00000
#> 300   32400  100.00000
#> 301  106200  316.22777
#> 302   57900  316.22777
#> 303   39900  316.22777
#> 304   36100  316.22777
#> 305   33800  316.22777
#> 306   33700  316.22777
#> 307   33200  316.22777
#> 308   32700  316.22777
#> 309   33000  316.22777
#> 310   32600  316.22777
#> 311   32600  316.22777
#> 312   33000  316.22777
#> 313   32700  316.22777
#> 314   32500  316.22777
#> 315   32600  316.22777
#> 316   32600  316.22777
#> 317   32300  316.22777
#> 318   32500  316.22777
#> 319   32000  316.22777
#> 320   32400  316.22777
#> 321   32300  316.22777
#> 322   32600  316.22777
#> 323   32400  316.22777
#> 324   32500  316.22777
#> 325   36200  316.22777
#> 326   32900  316.22777
#> 327   32300  316.22777
#> 328   32800  316.22777
#> 329   32200  316.22777
#> 330   32800  316.22777
#> 331   79900  316.22777
#> 332   48900  316.22777
#> 333   35700  316.22777
#> 334   33700  316.22777
#> 335   33400  316.22777
#> 336   33000  316.22777
#> 337   32600  316.22777
#> 338   33900  316.22777
#> 339   32800  316.22777
#> 340   33100  316.22777
#> 341   32700  316.22777
#> 342   33000  316.22777
#> 343   32300  316.22777
#> 344   32300  316.22777
#> 345   32500  316.22777
#> 346   32500  316.22777
#> 347   32400  316.22777
#> 348   32600  316.22777
#> 349   32400  316.22777
#> 350   34800  316.22777
#> 351   32900  316.22777
#> 352   32900  316.22777
#> 353   33500  316.22777
#> 354   33100  316.22777
#> 355   33100  316.22777
#> 356   33300  316.22777
#> 357   33000  316.22777
#> 358   33000  316.22777
#> 359   49900  316.22777
#> 360   40500  316.22777
#> 361   38300  316.22777
#> 362   38600  316.22777
#> 363   38100  316.22777
#> 364   38100  316.22777
#> 365   38200  316.22777
#> 366   39000  316.22777
#> 367   38000  316.22777
#> 368   38700  316.22777
#> 369   38900  316.22777
#> 370   39600  316.22777
#> 371   38800  316.22777
#> 372   38800  316.22777
#> 373   38500  316.22777
#> 374   38400  316.22777
#> 375   38300  316.22777
#> 376   39800  316.22777
#> 377   39700  316.22777
#> 378   39600  316.22777
#> 379   38500  316.22777
#> 380   38900  316.22777
#> 381   39000  316.22777
#> 382   38500  316.22777
#> 383   38800  316.22777
#> 384   44500  316.22777
#> 385   34200  316.22777
#> 386   33100  316.22777
#> 387   33500  316.22777
#> 388   33000  316.22777
#> 389   33900  316.22777
#> 390   32800  316.22777
#> 391   34100  316.22777
#> 392   32800  316.22777
#> 393   33600  316.22777
#> 394   32800  316.22777
#> 395   32500  316.22777
#> 396   32600  316.22777
#> 397   32500  316.22777
#> 398   32200  316.22777
#> 399   32500  316.22777
#> 400   32700  316.22777
#> 401  101500 1000.00000
#> 402   45200 1000.00000
#> 403   39100 1000.00000
#> 404   34900 1000.00000
#> 405   33100 1000.00000
#> 406   32500 1000.00000
#> 407   32300 1000.00000
#> 408   31500 1000.00000
#> 409   32000 1000.00000
#> 410   31800 1000.00000
#> 411   31700 1000.00000
#> 412   32200 1000.00000
#> 413   32000 1000.00000
#> 414   31400 1000.00000
#> 415   31800 1000.00000
#> 416   31500 1000.00000
#> 417   31400 1000.00000
#> 418   31800 1000.00000
#> 419   31200 1000.00000
#> 420   31400 1000.00000
#> 421   35900 1000.00000
#> 422   64500 1000.00000
#> 423   72200 1000.00000
#> 424   73300 1000.00000
#> 425   70800 1000.00000
#> 426   75600 1000.00000
#> 427   71600 1000.00000
#> 428   71300 1000.00000
#> 429   45200 1000.00000
#> 430   39600 1000.00000
#> 431   38700 1000.00000
#> 432   38100 1000.00000
#> 433   37800 1000.00000
#> 434   37300 1000.00000
#> 435   37100 1000.00000
#> 436   36700 1000.00000
#> 437   37100 1000.00000
#> 438   36900 1000.00000
#> 439   37400 1000.00000
#> 440   37300 1000.00000
#> 441   37300 1000.00000
#> 442   37700 1000.00000
#> 443   37600 1000.00000
#> 444   37600 1000.00000
#> 445   37300 1000.00000
#> 446   43600 1000.00000
#> 447   71100 1000.00000
#> 448   64300 1000.00000
#> 449   40100 1000.00000
#> 450   35000 1000.00000
#> 451   32800 1000.00000
#> 452   32100 1000.00000
#> 453   32200 1000.00000
#> 454   32700 1000.00000
#> 455   32000 1000.00000
#> 456   32100 1000.00000
#> 457   31600 1000.00000
#> 458   31700 1000.00000
#> 459   32200 1000.00000
#> 460   31900 1000.00000
#> 461   31300 1000.00000
#> 462   31600 1000.00000
#> 463   31200 1000.00000
#> 464   31300 1000.00000
#> 465   30900 1000.00000
#> 466   32700 1000.00000
#> 467   31600 1000.00000
#> 468   32000 1000.00000
#> 469   31600 1000.00000
#> 470   32500 1000.00000
#> 471   31500 1000.00000
#> 472   31100 1000.00000
#> 473   31300 1000.00000
#> 474   32000 1000.00000
#> 475   50100 1000.00000
#> 476   39900 1000.00000
#> 477   39000 1000.00000
#> 478   38600 1000.00000
#> 479   37100 1000.00000
#> 480   37700 1000.00000
#> 481   37700 1000.00000
#> 482   37500 1000.00000
#> 483   37800 1000.00000
#> 484   37400 1000.00000
#> 485   37200 1000.00000
#> 486   37500 1000.00000
#> 487   37100 1000.00000
#> 488   37500 1000.00000
#> 489   37500 1000.00000
#> 490   37100 1000.00000
#> 491   38200 1000.00000
#> 492   37400 1000.00000
#> 493   37800 1000.00000
#> 494   37100 1000.00000
#> 495   37500 1000.00000
#> 496   36800 1000.00000
#> 497   37000 1000.00000
#> 498   36800 1000.00000
#> 499   37200 1000.00000
#> 500   37200 1000.00000